@@ -1,15 +1,24 @@
 module bennypowers.dev/cem
 
-go 1.24.3
+go 1.25
 
 require (
+	atomicgo.dev/keyboard v0.2.9
 	github.com/IBM/fp-go v1.0.153
 	github.com/adrg/xdg v0.5.3
 	github.com/agext/levenshtein v1.2.3
+	github.com/alecthomas/chroma/v2 v2.20.0
 	github.com/bmatcuk/doublestar v1.3.4
+	github.com/bmatcuk/doublestar/v4 v4.9.1
 	github.com/dunglas/go-urlpattern v0.0.0-20241020164140-716dfa1c80b1
+	github.com/evanw/esbuild v0.27.0
 	github.com/fsnotify/fsnotify v1.9.0
+	github.com/google/go-cmp v0.7.0
+	github.com/google/jsonschema-go v0.2.1-0.20250825175020-748c325cec76
+	github.com/gorilla/websocket v1.5.3
 	github.com/gosimple/slug v1.15.0
+	github.com/microcosm-cc/bluemonday v1.0.27
+	github.com/modelcontextprotocol/go-sdk v0.4.0
 	github.com/nsf/jsondiff v0.0.0-20230430225905-43f6cf3098c1
 	github.com/pterm/pterm v0.12.81
 	github.com/sabhiram/go-gitignore v0.0.0-20210923224102-525f6e181f06
