@@ -27,6 +27,7 @@ import (
 
 	"github.com/pterm/pterm"
 
+	"bennypowers.dev/cem/generate/jsdoc"
 	M "bennypowers.dev/cem/manifest"
 	Q "bennypowers.dev/cem/queries"
 	ts "github.com/tree-sitter/go-tree-sitter"
@@ -192,11 +193,9 @@ func amendStylesMapFromSource(
 				pterm.Warning.Printf("%s:%d: Ambiguous comment ignored: more than one var() call in declaration.\n", path, line)
 			} else {
 				for _, comment := range comment {
-					info, err := NewCssCustomPropertyInfo(comment.Text, queryManager)
+					err := jsdoc.EnrichCSSPropertyWithJSDoc(comment.Text, &p, queryManager)
 					if err != nil {
 						errs = errors.Join(errs, err)
-					} else {
-						info.MergeToCssCustomProperty(&p)
 					}
 				}
 			}
@@ -229,7 +228,7 @@ func (mp *ModuleProcessor) processStyles(captures Q.CaptureMap) (props CssPropsM
 					} else {
 						content, err := os.ReadFile(absPath)
 						if err != nil {
-							errs = errors.Join(errs, fmt.Errorf(`Could not open css file
+							errs = errors.Join(errs, fmt.Errorf(`could not open css file
 	imported as %s
 	at path %s
 	from module directory %s: %w`, spec, absPath, moduleDir, err))
