@@ -28,7 +28,7 @@ import (
 	DD "bennypowers.dev/cem/generate/demodiscovery"
 	M "bennypowers.dev/cem/manifest"
 	Q "bennypowers.dev/cem/queries"
-	W "bennypowers.dev/cem/workspace"
+	"bennypowers.dev/cem/types"
 
 	DS "github.com/bmatcuk/doublestar"
 	ts "github.com/tree-sitter/go-tree-sitter"
@@ -51,13 +51,13 @@ func matchesAnyPattern(file string, patterns []string) bool {
 
 type preprocessResult struct {
 	demoFiles       []string
-	designTokens    *DT.DesignTokens
+	designTokens    types.DesignTokens
 	excludePatterns []string
 	includedFiles   []string
 }
 
 // preprocess handles config merging for generate command
-func preprocess(ctx W.WorkspaceContext) (r preprocessResult, errs error) {
+func preprocess(ctx types.WorkspaceContext) (r preprocessResult, errs error) {
 	cfg, err := ctx.Config()
 	if err != nil {
 		return r, err
@@ -73,11 +73,12 @@ func preprocess(ctx W.WorkspaceContext) (r preprocessResult, errs error) {
 	}
 
 	if cfg.Generate.DesignTokens.Spec != "" {
-		tokens, err := DT.LoadDesignTokens(ctx)
+		designTokens, err := validateAndLoadDesignTokens(ctx)
 		if err != nil {
 			errs = errors.Join(errs, err)
+		} else {
+			r.designTokens = designTokens
 		}
-		r.designTokens = tokens
 	}
 	if cfg.Generate.DemoDiscovery.FileGlob != "" {
 		demoFiles, err := ctx.Glob(cfg.Generate.DemoDiscovery.FileGlob)
@@ -106,7 +107,7 @@ func preprocess(ctx W.WorkspaceContext) (r preprocessResult, errs error) {
 
 type processJob struct {
 	file string
-	ctx  W.WorkspaceContext
+	ctx  types.WorkspaceContext
 }
 
 func processModule(
@@ -115,21 +116,21 @@ func processModule(
 	parser *ts.Parser,
 	depTracker *FileDependencyTracker,
 	cssCache CssCache,
-) (module *M.Module, tagAliases map[string]string, logCtx *LogCtx, errs error) {
+) (module *M.Module, tagAliases map[string]string, typeAliases map[string]string, imports map[string]importInfo, logCtx *LogCtx, errs error) {
 	defer parser.Reset()
 	mp, err := NewModuleProcessor(job.ctx, job.file, parser, qm, cssCache)
 	if err != nil {
-		return nil, nil, nil, err
+		return nil, nil, nil, nil, nil, err
 	}
 	defer mp.Close()
 	cfg, err := job.ctx.Config()
 	if err != nil {
-		return nil, nil, nil, err
+		return nil, nil, nil, nil, nil, err
 	}
 	if cfg.Verbose {
 		mp.logger.Section.Printf("Module: %s", mp.logger.File)
 	}
-	module, tagAliases, err = mp.Collect()
+	module, tagAliases, typeAliases, imports, err = mp.Collect()
 
 	// Record dependencies for incremental rebuilds
 	if depTracker != nil && module != nil {
@@ -149,13 +150,15 @@ func processModule(
 		}
 	}
 
-	return module, tagAliases, mp.logger, err
+	return module, tagAliases, typeAliases, imports, mp.logger, err
 }
 
 func postprocess(
-	ctx W.WorkspaceContext,
+	ctx types.WorkspaceContext,
 	result preprocessResult,
 	allTagAliases map[string]string,
+	typeAliases moduleTypeAliasesMap,
+	imports moduleImportsMap,
 	qm *Q.QueryManager,
 	modules []M.Module,
 ) (pkg M.Package, errs error) {
@@ -172,7 +175,7 @@ func postprocess(
 	if cfgErr == nil {
 		urlPattern = cfg.Generate.DemoDiscovery.URLPattern
 	}
-	demoMap, err := DD.NewDemoMapWithPattern(result.demoFiles, urlPattern, allTagAliases)
+	demoMap, err := DD.NewDemoMapWithPattern(ctx, result.demoFiles, urlPattern, allTagAliases)
 	if err != nil {
 		errsList = append(errsList, err)
 	}
@@ -183,7 +186,7 @@ func postprocess(
 		go func(module *M.Module) {
 			defer wg.Done()
 			if result.designTokens != nil {
-				DT.MergeDesignTokensToModule(module, *result.designTokens)
+				DT.MergeDesignTokensToModule(module, result.designTokens)
 			}
 			// Discover demos and attach to manifest
 			if len(demoMap) > 0 {
@@ -202,6 +205,10 @@ func postprocess(
 	slices.SortStableFunc(pkg.Modules, func(a, b M.Module) int {
 		return cmp.Compare(a.Path, b.Path)
 	})
+	// Resolve type aliases
+	if err := ResolveTypeAliases(&pkg, typeAliases, imports); err != nil {
+		errsList = append(errsList, fmt.Errorf("type resolution failed: %w", err))
+	}
 	if len(errsList) > 0 {
 		errs = errors.Join(errsList...)
 	}
@@ -209,7 +216,7 @@ func postprocess(
 }
 
 // Generates a custom-elements manifest from a list of typescript files
-func Generate(ctx W.WorkspaceContext) (manifest *string, errs error) {
+func Generate(ctx types.WorkspaceContext) (manifest *string, errs error) {
 	session, err := NewGenerateSession(ctx)
 	if err != nil {
 		return nil, err
@@ -226,4 +233,19 @@ func Generate(ctx W.WorkspaceContext) (manifest *string, errs error) {
 		return nil, fmt.Errorf("module serialize failed: %w", err)
 	}
 	return &manifestStr, nil
+}
+
+// validateAndLoadDesignTokens loads design tokens from cache.
+// Returns a proper error if the cached object cannot be loaded.
+func validateAndLoadDesignTokens(ctx types.WorkspaceContext) (types.DesignTokens, error) {
+	tokens, err := ctx.DesignTokensCache().LoadOrReuse(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to load design tokens: %w", err)
+	}
+
+	if tokens == nil {
+		return nil, fmt.Errorf("design tokens cache returned nil - check design tokens spec configuration")
+	}
+
+	return tokens, nil
 }
