@@ -18,14 +18,15 @@ package generate
 
 import (
 	"errors"
+	"fmt"
 
+	"bennypowers.dev/cem/generate/jsdoc"
 	M "bennypowers.dev/cem/manifest"
 	Q "bennypowers.dev/cem/queries"
 )
 
-func generateVarDeclaration(
+func (mp *ModuleProcessor) generateVarDeclaration(
 	captures Q.CaptureMap,
-	queryManager *Q.QueryManager,
 ) (declaration *M.VariableDeclaration, errs error) {
 	nameNodes, ok := captures["variable.name"]
 	if !ok || len(nameNodes) <= 0 {
@@ -43,6 +44,15 @@ func generateVarDeclaration(
 		},
 	}
 
+	// Add source reference if available
+	nameNode := Q.GetDescendantById(mp.root, nameNodes[0].NodeId)
+	sourceRef, sourceErr := mp.generateSourceReference(nameNode)
+	if sourceErr != nil {
+		errs = errors.Join(errs, fmt.Errorf("failed to generate source reference for variable %s: %w", varName, sourceErr))
+	} else {
+		declaration.Source = sourceRef
+	}
+
 	typeNodes, ok := captures["variable.type"]
 	if ok && len(typeNodes) > 0 {
 		declaration.Type = &M.Type{
@@ -50,13 +60,11 @@ func generateVarDeclaration(
 		}
 	}
 
-	jsdoc, ok := captures["variable.jsdoc"]
-	if ok && len(jsdoc) > 0 {
-		info, err := NewPropertyInfo(jsdoc[0].Text, queryManager)
+	jsdocNodes, ok := captures["variable.jsdoc"]
+	if ok && len(jsdocNodes) > 0 {
+		err := jsdoc.EnrichPropertyWithJSDoc(jsdocNodes[0].Text, &declaration.PropertyLike, mp.queryManager)
 		if err != nil {
 			errs = errors.Join(errs, err)
-		} else {
-			info.MergeToPropertyLike(&declaration.PropertyLike)
 		}
 	}
 
