@@ -93,6 +93,8 @@ type FSNotifyFileWatcher struct {
 	errors  chan error
 	mu      sync.RWMutex
 	closed  bool
+	done    chan struct{}  // Signal to stop translateEvents goroutine
+	wg      sync.WaitGroup // Wait for goroutine to exit
 }
 
 // NewFSNotifyFileWatcher creates a new file watcher using fsnotify.
@@ -106,10 +108,15 @@ func NewFSNotifyFileWatcher() (*FSNotifyFileWatcher, error) {
 		watcher: watcher,
 		events:  make(chan FileWatchEvent, 100),
 		errors:  make(chan error, 10),
+		done:    make(chan struct{}),
 	}
 
 	// Start event translation goroutine
-	go fw.translateEvents()
+	fw.wg.Add(1)
+	go func() {
+		defer fw.wg.Done()
+		fw.translateEvents()
+	}()
 
 	return fw, nil
 }
@@ -138,13 +145,23 @@ func (fw *FSNotifyFileWatcher) Remove(name string) error {
 
 func (fw *FSNotifyFileWatcher) Close() error {
 	fw.mu.Lock()
-	defer fw.mu.Unlock()
 
 	if fw.closed {
+		fw.mu.Unlock()
 		return nil
 	}
 
 	fw.closed = true
+
+	// Signal the translateEvents goroutine to stop
+	close(fw.done)
+
+	fw.mu.Unlock()
+
+	// Wait for the goroutine to exit
+	fw.wg.Wait()
+
+	// Now close the watcher and channels
 	err := fw.watcher.Close()
 	close(fw.events)
 	close(fw.errors)
@@ -162,6 +179,14 @@ func (fw *FSNotifyFileWatcher) Errors() <-chan error {
 
 // translateEvents converts fsnotify events to our abstracted events
 func (fw *FSNotifyFileWatcher) translateEvents() {
+	// Add an initial blocking operation to prevent CPU spinning
+	// Wait a moment for the watcher to be fully initialized
+	select {
+	case <-fw.done:
+		return
+	default:
+	}
+
 	for {
 		select {
 		case event, ok := <-fw.watcher.Events:
@@ -189,9 +214,14 @@ func (fw *FSNotifyFileWatcher) translateEvents() {
 
 			fw.mu.RLock()
 			if !fw.closed {
-				fw.events <- FileWatchEvent{
+				select {
+				case fw.events <- FileWatchEvent{
 					Name: event.Name,
 					Op:   op,
+				}:
+				case <-fw.done:
+					fw.mu.RUnlock()
+					return
 				}
 			}
 			fw.mu.RUnlock()
@@ -203,9 +233,32 @@ func (fw *FSNotifyFileWatcher) translateEvents() {
 
 			fw.mu.RLock()
 			if !fw.closed {
-				fw.errors <- err
+				select {
+				case fw.errors <- err:
+				case <-fw.done:
+					fw.mu.RUnlock()
+					return
+				}
 			}
 			fw.mu.RUnlock()
+
+		case <-fw.done:
+			// Shutdown signal received
+			return
 		}
 	}
+}
+
+// GenerateWatcher provides an abstraction over generate watching operations.
+// This interface enables:
+// - Testing with mock generate watchers (instant callbacks)
+// - In-process vs subprocess implementations
+// - Graceful handling of generate watcher unavailability
+type GenerateWatcher interface {
+	// Start begins watching for source file changes and generating manifests
+	Start() error
+	// Stop ceases watching and cleans up resources
+	Stop() error
+	// IsRunning returns whether the watcher is currently active
+	IsRunning() bool
 }
