@@ -1,6 +1,3 @@
-//go:build race_unsafe
-// +build race_unsafe
-
 /*
 Copyright © 2025 Benny Powers <web@bennypowers.com>
 
@@ -18,120 +15,108 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-// This file contains tests that trigger Go's race detector "hole in findfunctab" issue.
+// This file contains tests for concurrent operations using Go 1.25's testing/synctest package.
 //
-// These tests are excluded from normal runs due to a known limitation in Go's race detector
-// that occurs with certain channel operations. The issue causes false failures with
-// race detection even when no actual race conditions exist.
+// These tests previously had issues with Go's race detector "hole in findfunctab" limitation,
+// but Go 1.25's synctest package provides proper isolation for concurrent testing scenarios.
 //
-// ROOT CAUSE:
-// - MockTimeProvider.After() returns <-chan time.Time (creates and closes channels)
-// - MockFileWatcher.Events() returns <-chan FileWatchEvent (creates channels)
-// - FSNotifyFileWatcher also creates channels in the same package
-// - Go's race detector has issues with certain channel creation/closure patterns
+// The synctest package provides:
+// - Isolated "bubble" environments for concurrent tests
+// - Virtual time that advances when goroutines block
+// - Proper synchronization primitives for testing concurrent code
 //
-// TO RUN THESE TESTS:
-//   go test -tags=race_unsafe ./internal/platform/
-//
-// STATUS:
-// - Known Go toolchain limitation, not our code
-// - No timeline for fix from Go team
-// - Tests work correctly, only race detector has issues
-// - The mocks work correctly in production use
-//
-// WORKAROUND:
-// - Tests excluded by default to allow race detection on main codebase
-// - Can be run manually with build tags for verification
-// - Integration tests in other packages verify the mocks work correctly
-//
-// See: https://github.com/golang/go/issues/[race-detector-hole-in-findfunctab]
+// This resolves the previous race detector issues with channel operations in mocks.
 
 package platform_test
 
 import (
 	"testing"
+	"testing/synctest"
 	"time"
 
 	"bennypowers.dev/cem/internal/platform"
 )
 
 func TestMockTimeProvider(t *testing.T) {
-	startTime := time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC)
-	mockTime := platform.NewMockTimeProvider(startTime)
+	synctest.Test(t, func(t *testing.T) {
+		startTime := time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC)
+		mockTime := platform.NewMockTimeProvider(startTime)
 
-	// Test initial time
-	if mockTime.Now() != startTime {
-		t.Errorf("Expected initial time %v, got %v", startTime, mockTime.Now())
-	}
-
-	// Test Sleep - should advance time instantly
-	mockTime.Sleep(5 * time.Second)
-	expectedTime := startTime.Add(5 * time.Second)
-	if mockTime.Now() != expectedTime {
-		t.Errorf("Expected time after sleep %v, got %v", expectedTime, mockTime.Now())
-	}
-
-	// Test sleep calls tracking
-	sleepCalls := mockTime.GetSleepCalls()
-	if len(sleepCalls) != 1 || sleepCalls[0] != 5*time.Second {
-		t.Errorf("Expected sleep calls [5s], got %v", sleepCalls)
-	}
-
-	// Test After channel - this triggers the race detector issue
-	ch := mockTime.After(1 * time.Second)
-	select {
-	case receivedTime := <-ch:
-		expectedAfterTime := expectedTime.Add(1 * time.Second)
-		if receivedTime != expectedAfterTime {
-			t.Errorf("Expected After time %v, got %v", expectedAfterTime, receivedTime)
+		// Test initial time
+		if mockTime.Now() != startTime {
+			t.Errorf("Expected initial time %v, got %v", startTime, mockTime.Now())
 		}
-	default:
-		t.Error("After channel should have delivered time immediately")
-	}
+
+		// Test Sleep - should advance time instantly
+		mockTime.Sleep(5 * time.Second)
+		expectedTime := startTime.Add(5 * time.Second)
+		if mockTime.Now() != expectedTime {
+			t.Errorf("Expected time after sleep %v, got %v", expectedTime, mockTime.Now())
+		}
+
+		// Test sleep calls tracking
+		sleepCalls := mockTime.GetSleepCalls()
+		if len(sleepCalls) != 1 || sleepCalls[0] != 5*time.Second {
+			t.Errorf("Expected sleep calls [5s], got %v", sleepCalls)
+		}
+
+		// Test After channel - now works correctly with synctest isolation
+		ch := mockTime.After(1 * time.Second)
+		select {
+		case receivedTime := <-ch:
+			expectedAfterTime := expectedTime.Add(1 * time.Second)
+			if receivedTime != expectedAfterTime {
+				t.Errorf("Expected After time %v, got %v", expectedAfterTime, receivedTime)
+			}
+		default:
+			t.Error("After channel should have delivered time immediately")
+		}
+	})
 }
 
 func TestMockFileWatcher(t *testing.T) {
-	watcher := platform.NewMockFileWatcher()
-	defer watcher.Close()
+	synctest.Test(t, func(t *testing.T) {
+		watcher := platform.NewMockFileWatcher()
+		defer func() { _ = watcher.Close() }()
 
-	// Test adding watch paths
-	err := watcher.Add("/test/path")
-	if err != nil {
-		t.Fatalf("Failed to add watch path: %v", err)
-	}
-
-	// Verify path is watched
-	watchedPaths := watcher.GetWatchedPaths()
-	if len(watchedPaths) != 1 || watchedPaths[0] != "/test/path" {
-		t.Errorf("Expected watched paths [/test/path], got %v", watchedPaths)
-	}
-
-	// Test removing watch paths
-	err = watcher.Remove("/test/path")
-	if err != nil {
-		t.Fatalf("Failed to remove watch path: %v", err)
-	}
-
-	watchedPaths = watcher.GetWatchedPaths()
-	if len(watchedPaths) != 0 {
-		t.Errorf("Expected no watched paths, got %v", watchedPaths)
-	}
-
-	// Test triggering events - this triggers the race detector issue
-	watcher.Add("/test/path") // Re-add for event testing
-	watcher.TriggerEvent("/test/path/file.txt", platform.Write)
-
-	// Attempt to read from Events() channel - this causes race detector issues
-	select {
-	case event := <-watcher.Events():
-		if event.Name != "/test/path/file.txt" {
-			t.Errorf("Expected event name /test/path/file.txt, got %s", event.Name)
+		// Test adding watch paths
+		err := watcher.Add("/test/path")
+		if err != nil {
+			t.Fatalf("Failed to add watch path: %v", err)
 		}
-		if event.Op != platform.Write {
-			t.Errorf("Expected Write operation, got %v", event.Op)
+
+		// Verify path is watched
+		watchedPaths := watcher.GetWatchedPaths()
+		if len(watchedPaths) != 1 || watchedPaths[0] != "/test/path" {
+			t.Errorf("Expected watched paths [/test/path], got %v", watchedPaths)
 		}
-	default:
-		// This is expected when race detector issues prevent channel operations
-		t.Logf("No event received - this may be due to race detector channel issues")
-	}
+
+		// Test removing watch paths
+		err = watcher.Remove("/test/path")
+		if err != nil {
+			t.Fatalf("Failed to remove watch path: %v", err)
+		}
+
+		watchedPaths = watcher.GetWatchedPaths()
+		if len(watchedPaths) != 0 {
+			t.Errorf("Expected no watched paths, got %v", watchedPaths)
+		}
+
+		// Test triggering events - now works correctly with synctest isolation
+		_ = watcher.Add("/test/path") // Re-add for event testing
+		watcher.TriggerEvent("/test/path/file.txt", platform.Write)
+
+		// Read from Events() channel - synctest provides proper isolation
+		select {
+		case event := <-watcher.Events():
+			if event.Name != "/test/path/file.txt" {
+				t.Errorf("Expected event name /test/path/file.txt, got %s", event.Name)
+			}
+			if event.Op != platform.Write {
+				t.Errorf("Expected Write operation, got %v", event.Op)
+			}
+		default:
+			t.Error("Expected to receive file watcher event")
+		}
+	})
 }
