@@ -23,7 +23,7 @@ import (
 	"strings"
 	"testing"
 
-	"bennypowers.dev/cem/lsp"
+	"bennypowers.dev/cem/lsp/document"
 	"bennypowers.dev/cem/lsp/testhelpers"
 	"bennypowers.dev/cem/lsp/types"
 	protocol "github.com/tliron/glsp/protocol_3_16"
@@ -67,7 +67,7 @@ func TestIndentationDetection(t *testing.T) {
 			ctx := testhelpers.NewMockServerContext()
 
 			// Create DocumentManager and add document
-			dm, err := lsp.NewDocumentManager()
+			dm, err := document.NewDocumentManager()
 			if err != nil {
 				t.Fatalf("Failed to create DocumentManager: %v", err)
 			}
