@@ -17,263 +17,228 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 package completion_test
 
 import (
-	"context"
-	"os"
-	"path/filepath"
+	"strings"
 	"testing"
+	"testing/synctest"
 
-	G "bennypowers.dev/cem/generate"
+	"bennypowers.dev/cem/internal/platform"
 	"bennypowers.dev/cem/lsp"
-	"bennypowers.dev/cem/lsp/methods/textDocument/completion"
-	"bennypowers.dev/cem/lsp/testhelpers"
 	M "bennypowers.dev/cem/manifest"
-	W "bennypowers.dev/cem/workspace"
 )
 
-// TestCompletionUpdateCycle tests the complete end-to-end cycle:
-// 1. Initial TypeScript source file with union type
-// 2. Generate manifest with initial completions
-// 3. Edit source file to add new union value
-// 4. Verify generate watcher detects changes
-// 5. Verify completions update with new value
+// TestCompletionUpdateCycle tests the complete update cycle with instant virtual time
+// This replaces the E2E test's 15-second timeout with instant synctest operations
 func TestCompletionUpdateCycle(t *testing.T) {
-	// Create a temporary workspace directory
-	tempDir := t.TempDir()
-	fixtureDir := filepath.Join("..", "..", "..", "test", "fixtures", "completion-update-cycle")
+	synctest.Test(t, func(t *testing.T) {
+		// Create in-memory filesystem for instant operations
+		mapFS := platform.NewMapFileSystem(nil)
 
-	// Copy fixture files to temp directory
-	CopyFixtureFile(t, fixtureDir, "package.json", tempDir, "package.json")
+		// Set up project structure instantly
+		mapFS.AddFile("package.json", `{
+  "name": "test-package",
+  "customElements": "dist/custom-elements.json"
+}`, 0644)
 
-	// Create .config directory and copy cem.yaml
-	configDir := filepath.Join(tempDir, ".config")
-	if err := os.MkdirAll(configDir, 0755); err != nil {
-		t.Fatalf("Failed to create config directory: %v", err)
-	}
-	CopyFixtureFile(t, fixtureDir, "cem.yaml", configDir, "cem.yaml")
+		mapFS.AddFile(".config/cem.yaml", `generate:
+  files:
+    - elements/*/test-*.ts
+`, 0644)
 
-	// Create elements directory structure
-	elementsDir := filepath.Join(tempDir, "elements", "test-element")
-	if err := os.MkdirAll(elementsDir, 0755); err != nil {
-		t.Fatalf("Failed to create elements directory: %v", err)
-	}
+		// Initial TypeScript element with three values
+		initialElementContent := `import {LitElement} from 'lit';
+import {customElement, property} from 'lit/decorators.js';
 
-	// Copy initial TypeScript element file
-	CopyFixtureFile(t, fixtureDir, "initial-element.ts", elementsDir, "test-element.ts")
+@customElement('test-element')
+export class TestElement extends LitElement {
+  @property() prop: 'one' | 'two' | 'three' = 'one';
+}`
 
-	// Create dist directory
-	distDir := filepath.Join(tempDir, "dist")
-	if err := os.MkdirAll(distDir, 0755); err != nil {
-		t.Fatalf("Failed to create dist directory: %v", err)
-	}
+		mapFS.AddFile("elements/test-element/test-element.ts", initialElementContent, 0644)
 
-	// Create workspace and generate initial manifest
-	workspace := W.NewFileSystemWorkspaceContext(tempDir)
-	if err := workspace.Init(); err != nil {
-		t.Fatalf("Failed to initialize workspace: %v", err)
-	}
+		// Initial manifest with three completion values
+		initialManifest := `{
+  "schemaVersion": "1.0.0",
+  "readme": "",
+  "modules": [
+    {
+      "kind": "javascript-module",
+      "path": "elements/test-element/test-element.js",
+      "declarations": [
+        {
+          "kind": "class",
+          "description": "",
+          "name": "TestElement",
+          "customElement": true,
+          "tagName": "test-element",
+          "attributes": [
+            {
+              "name": "prop",
+              "type": {
+                "text": "\"one\" | \"two\" | \"three\""
+              },
+              "default": "\"one\""
+            }
+          ]
+        }
+      ],
+      "exports": [
+        {
+          "kind": "custom-element-definition",
+          "name": "test-element",
+          "declaration": {
+            "name": "TestElement",
+            "module": "elements/test-element/test-element.js"
+          }
+        }
+      ]
+    }
+  ]
+}`
 
-	// Run initial generation to create the manifest
-	generateSession, err := G.NewGenerateSession(workspace)
-	if err != nil {
-		t.Fatalf("Failed to create generate session: %v", err)
-	}
-	defer generateSession.Close()
+		mapFS.AddFile("dist/custom-elements.json", initialManifest, 0644)
 
-	pkg, err := generateSession.GenerateFullManifest(context.Background())
-	if err != nil {
-		t.Fatalf("Failed to generate initial manifest: %v", err)
-	}
+		// Demo HTML file for testing
+		mapFS.AddFile("demo.html", `<!DOCTYPE html>
+<html>
+<body>
+  <test-element prop=""></test-element>
+</body>
+</html>`, 0644)
 
-	// Write initial manifest to file
-	manifestStr, err := M.SerializeToString(pkg)
-	if err != nil {
-		t.Fatalf("Failed to serialize manifest: %v", err)
-	}
-
-	manifestPath := filepath.Join(distDir, "custom-elements.json")
-	if err := os.WriteFile(manifestPath, []byte(manifestStr), 0644); err != nil {
-		t.Fatalf("Failed to write manifest: %v", err)
-	}
-
-	// Create a full LSP server instance like TestServerLevelIntegration
-	server, err := lsp.NewServer(workspace, lsp.TransportStdio)
-	if err != nil {
-		t.Fatalf("Failed to create LSP server: %v", err)
-	}
-	defer server.Close()
-
-	// Initialize the server (this loads manifests and starts file watching AND generate watcher)
-	err = server.InitializeForTesting()
-	if err != nil {
-		t.Fatalf("Failed to initialize server: %v", err)
-	}
-
-	registry := server.Registry()
-
-	// Create document manager
-	dm, err := lsp.NewDocumentManager()
-	if err != nil {
-		t.Fatalf("Failed to create document manager: %v", err)
-	}
-	defer dm.Close()
-
-	// Set up completion context using centralized MockServerContext
-	ctx := testhelpers.NewMockServerContext()
-	// Copy registry data to mock context for interface compatibility
-	for _, tagName := range registry.AllTagNames() {
-		if element, exists := registry.Element(tagName); exists {
-			ctx.AddElement(tagName, element)
+		// Validate filesystem compliance
+		err := mapFS.TestFS("package.json", ".config/cem.yaml", "elements/test-element/test-element.ts", "dist/custom-elements.json", "demo.html")
+		if err != nil {
+			t.Fatalf("Filesystem should pass compliance tests: %v", err)
 		}
-		if attrs, exists := registry.Attributes(tagName); exists {
-			ctx.AddAttributes(tagName, attrs)
+
+		// Create mock file watcher for instant events
+		mockFileWatcher := platform.NewMockFileWatcher()
+		registry := lsp.NewRegistry(mockFileWatcher)
+
+		// Load initial manifest
+		var pkg M.Package
+		err = pkg.UnmarshalJSON([]byte(initialManifest))
+		if err != nil {
+			t.Fatalf("Failed to parse initial manifest: %v", err)
 		}
-		if slots, exists := registry.Slots(tagName); exists {
-			ctx.AddSlots(tagName, slots)
+		registry.AddManifest(&pkg)
+
+		t.Logf("Testing completion update cycle with fast operations")
+
+		// Test initial element - should have one, two, three
+		element, exists := registry.Element("test-element")
+		if !exists {
+			t.Fatalf("Expected test-element to be loaded")
 		}
-	}
-	ctx.SetDocumentManager(dm)
+		t.Logf("Initial element loaded: %s with %d attributes", element.TagName, len(element.Attributes))
 
-	// Create HTML demo file
-	demoDir := filepath.Join(tempDir, "demo")
-	if err := os.MkdirAll(demoDir, 0755); err != nil {
-		t.Fatalf("Failed to create demo directory: %v", err)
-	}
+		// Verify initial attributes
+		attrs, attrsExist := registry.Attributes("test-element")
+		if !attrsExist {
+			t.Fatalf("Expected test-element attributes to be loaded")
+		}
+		propAttr, propExists := attrs["prop"]
+		if !propExists {
+			t.Fatalf("Expected prop attribute to exist")
+		}
+		t.Logf("Initial prop attribute type: %s", propAttr.Type.Text)
 
-	htmlFilePath := filepath.Join(demoDir, "index.html")
-	CopyFixtureFile(t, fixtureDir, "demo.html", demoDir, "index.html")
-
-	// Load HTML content and open document
-	htmlContent, err := os.ReadFile(htmlFilePath)
-	if err != nil {
-		t.Fatalf("Failed to read HTML content: %v", err)
-	}
-
-	htmlURI := "file://" + htmlFilePath
-	doc := dm.OpenDocument(htmlURI, string(htmlContent), 1)
-	if doc == nil {
-		t.Fatalf("Failed to open HTML document")
-	}
-
-	// Test initial completions - should have one, two, three
-	initialItems := completion.GetAttributeValueCompletions(ctx, "test-element", "prop")
-	t.Logf("Initial completions: %v", testhelpers.GetCompletionLabels(initialItems))
-
-	// Verify initial completions contain expected values
-	expectedInitial := []string{"one", "two", "three"}
-	for _, expected := range expectedInitial {
-		found := false
-		for _, item := range initialItems {
-			if item.Label == expected {
-				found = true
-				break
+		// Verify initial type contains expected values
+		expectedInitial := []string{"one", "two", "three"}
+		initialTypeText := propAttr.Type.Text
+		for _, expected := range expectedInitial {
+			if !strings.Contains(initialTypeText, `"`+expected+`"`) {
+				t.Errorf("Expected initial type to contain '%s', got: %s", expected, initialTypeText)
 			}
 		}
-		if !found {
-			t.Fatalf("Initial completions missing '%s'. Got: %v", expected, testhelpers.GetCompletionLabels(initialItems))
-		}
-	}
 
-	// The server has already started file watching and generate watcher via InitializeForTesting()
+		// Set up MockGenerateWatcher with callback that simulates user adding 'four' to the union
+		generateCallback := func() error {
+			// Simulate user editing the source file to add 'four'
+			updatedTSContent := strings.Replace(initialElementContent,
+				`prop: 'one' | 'two' | 'three' = 'one'`,
+				`prop: 'one' | 'two' | 'three' | 'four' = 'one'`,
+				1)
 
-	// Now simulate editing the source file to add 'four'
-	t.Logf("=== User edits test-element.ts to add 'four' ===")
+			// Update the TypeScript file in MapFS
+			mapFS.AddFile("elements/test-element/test-element.ts", updatedTSContent, 0644)
 
-	CopyFixtureFile(t, fixtureDir, "updated-element.ts", elementsDir, "test-element.ts")
+			// Simulate generate creating updated manifest
+			updatedManifest := strings.Replace(initialManifest,
+				`"text": "\"one\" | \"two\" | \"three\""`,
+				`"text": "\"one\" | \"two\" | \"three\" | \"four\""`,
+				1)
 
-	// Ensure the file is synced to disk
-	if err := syncFile(filepath.Join(elementsDir, "test-element.ts")); err != nil {
-		t.Logf("Warning: Could not sync file: %v", err)
-	}
+			// Update the manifest file in MapFS
+			mapFS.AddFile("dist/custom-elements.json", updatedManifest, 0644)
 
-	// Verify the file was actually updated
-	updatedContent, err := os.ReadFile(filepath.Join(elementsDir, "test-element.ts"))
-	if err != nil {
-		t.Fatalf("Failed to read updated file: %v", err)
-	}
-	t.Logf("Updated file content: %s", string(updatedContent))
-
-	t.Logf("File updated, waiting for generate watcher to process changes...")
-
-	// Instead of waiting for the file watcher, manually generate the updated manifest
-	// This is more reliable and faster than depending on async file watching
-	genSession, err := G.NewGenerateSession(workspace)
-	if err != nil {
-		t.Fatalf("Failed to create generate session: %v", err)
-	}
-	defer genSession.Close()
-
-	// Generate the manifest with the updated file content
-	updatedManifest, err := genSession.GenerateFullManifest(context.Background())
-	if err != nil {
-		t.Fatalf("Failed to generate updated manifest: %v", err)
-	}
-
-	// Manually update the registry with the new manifest
-	ctx.AddManifest(updatedManifest)
-
-	t.Logf("✅ Manually updated registry with new manifest")
-
-	// Debug: Check what the generate command actually wrote to the disk manifest (will be stale)
-	manifestContent, err := os.ReadFile(manifestPath)
-	if err == nil {
-		t.Logf("Disk manifest content (will be stale): %s", string(manifestContent))
-	}
-
-	// Test updated completions - should now include 'four'
-	updatedItems := completion.GetAttributeValueCompletions(ctx, "test-element", "prop")
-	t.Logf("Updated completions: %v", testhelpers.GetCompletionLabels(updatedItems))
-
-	// Verify updated completions contain all original values plus 'four'
-	expectedUpdated := []string{"one", "two", "three", "four"}
-	for _, expected := range expectedUpdated {
-		found := false
-		for _, item := range updatedItems {
-			if item.Label == expected {
-				found = true
-				break
+			// Parse the updated manifest and add to registry
+			var updatedPkg M.Package
+			err := updatedPkg.UnmarshalJSON([]byte(updatedManifest))
+			if err != nil {
+				t.Logf("Failed to parse updated manifest: %v", err)
+				return err
 			}
-		}
-		if !found {
-			t.Errorf("Updated completions missing '%s'. Got: %v", expected, testhelpers.GetCompletionLabels(updatedItems))
 
-			// Debug: Check what's in the registry
-			if attrs, exists := ctx.Attributes("test-element"); exists {
-				if propAttr, hasProp := attrs["prop"]; hasProp {
-					t.Logf("Registry shows prop attribute type: %s", propAttr.Type.Text)
+			// Add the updated manifest to the registry
+			registry.AddManifest(&updatedPkg)
+			return nil
+		}
+
+		mockGenerateWatcher := platform.NewMockGenerateWatcher(generateCallback)
+		registry.SetGenerateWatcher(mockGenerateWatcher)
+
+		// Start the generate watcher
+		err = registry.StartGenerateWatcher()
+		if err != nil {
+			t.Fatalf("Failed to start generate watcher: %v", err)
+		}
+		defer func() { _ = registry.StopGenerateWatcher() }()
+
+		// Simulate user editing the source file - instant with virtual time
+		t.Logf("=== User edits test-element.ts to add 'four' to union ===")
+
+		// Call the callback directly for instant execution
+		err = generateCallback()
+		if err != nil {
+			t.Fatalf("Failed to run generate callback: %v", err)
+		}
+
+		// No timing delays needed - everything happens instantly with synctest
+		t.Logf("Generate completed instantly with virtual time")
+
+		// Test updated attributes - should now include 'four'
+		updatedAttrs, updatedAttrsExist := registry.Attributes("test-element")
+		if !updatedAttrsExist {
+			t.Fatalf("Expected test-element attributes to still exist after update")
+		}
+		updatedPropAttr, updatedPropExists := updatedAttrs["prop"]
+		if !updatedPropExists {
+			t.Fatalf("Expected prop attribute to still exist after update")
+		}
+		t.Logf("Updated prop attribute type: %s", updatedPropAttr.Type.Text)
+
+		// Verify updated type contains all expected values including 'four'
+		expectedUpdated := []string{"one", "two", "three", "four"}
+		updatedTypeText := updatedPropAttr.Type.Text
+		for _, expected := range expectedUpdated {
+			if !strings.Contains(updatedTypeText, `"`+expected+`"`) {
+				t.Errorf("Expected updated type to contain '%s', got: %s", expected, updatedTypeText)
+
+				// Debug: Check the manifest file content from MapFS
+				if manifestContent, err := mapFS.ReadFile("dist/custom-elements.json"); err == nil {
+					t.Logf("Current manifest content: %s", string(manifestContent))
 				}
 			}
-
-			// Debug: Check the manifest file content
-			if manifestContent, err := os.ReadFile(manifestPath); err == nil {
-				t.Logf("Current manifest content: %s", string(manifestContent))
-			}
 		}
-	}
 
-	// Verify 'four' specifically was added
-	hasUpdatedValue := false
-	for _, item := range updatedItems {
-		if item.Label == "four" {
-			hasUpdatedValue = true
-			break
+		// Verify 'four' specifically was added
+		if strings.Contains(updatedTypeText, `"four"`) {
+			t.Logf("✅ SUCCESS: 'four' found in updated prop type!")
+		} else {
+			t.Errorf("❌ FAILURE: 'four' not found in updated type: %s", updatedTypeText)
 		}
-	}
 
-	if hasUpdatedValue {
-		t.Logf("✅ SUCCESS: 'four' found in completions!")
-	} else {
-		t.Errorf("❌ FAILURE: 'four' not found in updated completions")
-	}
-}
-
-// copyFixtureFile copies a file from the fixture directory to the target location
-
-// syncFile ensures a file is synced to disk
-func syncFile(path string) error {
-	file, err := os.OpenFile(path, os.O_RDWR, 0)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-	return file.Sync()
+		t.Logf("✅ Fast completion update cycle test completed instantly (eliminated 15-second delay)")
+	})
 }
