@@ -1,3 +1,19 @@
+/*
+Copyright © 2025 Benny Powers <web@bennypowers.com>
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
 package completion_test
 
 import (
@@ -8,7 +24,7 @@ import (
 	"strings"
 	"testing"
 
-	"bennypowers.dev/cem/lsp"
+	"bennypowers.dev/cem/lsp/document"
 	"bennypowers.dev/cem/lsp/methods/textDocument/completion"
 	"bennypowers.dev/cem/lsp/types"
 	protocol "github.com/tliron/glsp/protocol_3_16"
@@ -22,7 +38,7 @@ type TestHelpers struct{}
 // DEPRECATED: Tests should use dm.OpenDocument() directly
 func NewMockTemplateDocument(content, templateContext string) types.Document {
 	// Create a DocumentManager and use it to open the document
-	dm, err := lsp.NewDocumentManager()
+	dm, err := document.NewDocumentManager()
 	if err != nil {
 		panic("Failed to create DocumentManager for test: " + err.Error())
 	}
@@ -88,23 +104,6 @@ func (h *TestHelpers) GetAttributeValueCompletionsUsingMainEntry(
 	return completion.GetAttributeValueCompletions(ctx, tagName, attributeName), nil
 }
 
-// getDocumentManagerFromContext extracts DocumentManager from a completion context
-// This uses type assertion to check if the context has a document manager
-func getDocumentManagerFromContext(ctx types.ServerContext) *lsp.DocumentManager {
-	// Try to extract document manager using the ServerContext interface
-	dm, err := ctx.DocumentManager()
-	if err != nil {
-		return nil
-	}
-
-	// Type assert to concrete DocumentManager type
-	if lspDM, ok := dm.(*lsp.DocumentManager); ok {
-		return lspDM
-	}
-	return nil
-}
-
-// Global helper instance
 var TestHelper = &TestHelpers{}
 
 // MockDocument implements types.Document for testing
@@ -163,7 +162,7 @@ func CopyFixtureFiles(t *testing.T, fixtureDir, targetDir string) {
 		if err != nil {
 			return err
 		}
-		defer src.Close()
+		defer func() { _ = src.Close() }()
 
 		err = os.MkdirAll(filepath.Dir(targetPath), 0755)
 		if err != nil {
@@ -174,7 +173,7 @@ func CopyFixtureFiles(t *testing.T, fixtureDir, targetDir string) {
 		if err != nil {
 			return err
 		}
-		defer dst.Close()
+		defer func() { _ = dst.Close() }()
 
 		_, err = io.Copy(dst, src)
 		return err
@@ -196,7 +195,7 @@ func CopyFixtureFile(t *testing.T, fixtureDir, filename, targetDir, targetFilena
 	if err != nil {
 		t.Fatalf("Failed to open source file %s: %v", src, err)
 	}
-	defer srcFile.Close()
+	defer func() { _ = srcFile.Close() }()
 
 	err = os.MkdirAll(filepath.Dir(dst), 0755)
 	if err != nil {
@@ -207,7 +206,7 @@ func CopyFixtureFile(t *testing.T, fixtureDir, filename, targetDir, targetFilena
 	if err != nil {
 		t.Fatalf("Failed to create target file %s: %v", dst, err)
 	}
-	defer dstFile.Close()
+	defer func() { _ = dstFile.Close() }()
 
 	_, err = io.Copy(dstFile, srcFile)
 	if err != nil {
