@@ -17,9 +17,9 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 package references
 
 import (
-	"fmt"
-	"os"
+	"io/fs"
 	"path/filepath"
+	"strings"
 
 	"bennypowers.dev/cem/internal/platform"
 	"bennypowers.dev/cem/lsp/helpers"
@@ -63,7 +63,8 @@ func References(ctx types.ServerContext, context *glsp.Context, params *protocol
 	helpers.SafeDebugLog("[REFERENCES] Found element: %s", request.ElementName)
 
 	// Search for all references across all documents
-	return findAllReferences(ctx, request), nil
+	filesystem := ctx.FileSystem()
+	return findAllReferences(ctx, request, filesystem), nil
 }
 
 // ReferenceRequest contains information about what to search for
@@ -225,7 +226,7 @@ func containsHyphen(s string) bool {
 }
 
 // findAllReferences searches for all occurrences of the element across workspace
-func findAllReferences(ctx types.ServerContext, request *ReferenceRequest) []protocol.Location {
+func findAllReferences(ctx types.ServerContext, request *ReferenceRequest, filesystem platform.FileSystem) []protocol.Location {
 	var locations []protocol.Location
 
 	// First, search in all tracked (open) documents
@@ -240,7 +241,7 @@ func findAllReferences(ctx types.ServerContext, request *ReferenceRequest) []pro
 	// Also search workspace files that aren't currently open
 	workspaceRoot := ctx.WorkspaceRoot()
 	if workspaceRoot != "" {
-		workspaceLocations := findReferencesInWorkspace(workspaceRoot, request.ElementName, allDocuments)
+		workspaceLocations := findReferencesInWorkspaceWithFS(workspaceRoot, request.ElementName, allDocuments, filesystem)
 		locations = append(locations, workspaceLocations...)
 	}
 
@@ -284,9 +285,13 @@ func findReferencesInDocument(ctx types.ServerContext, doc types.Document, eleme
 	return locations
 }
 
-// findReferencesInWorkspace searches for references in all workspace files
-func findReferencesInWorkspace(workspaceRoot string, elementName string, openDocuments []types.Document) []protocol.Location {
+// findReferencesInWorkspaceWithFS searches for references using a provided filesystem
+// This allows for testability with mock filesystems
+func findReferencesInWorkspaceWithFS(workspaceRoot string, elementName string, openDocuments []types.Document, filesystem platform.FileSystem) []protocol.Location {
 	var locations []protocol.Location
+
+	// Normalize workspace root - remove trailing slashes for consistent path handling
+	workspaceRoot = strings.TrimSuffix(workspaceRoot, "/")
 
 	// Create a set of open document URIs to avoid duplicates
 	openFiles := make(map[string]bool)
@@ -297,12 +302,13 @@ func findReferencesInWorkspace(workspaceRoot string, elementName string, openDoc
 	// Load .gitignore if it exists
 	var gitignoreMatcher *ignore.GitIgnore
 	gitignorePath := filepath.Join(workspaceRoot, ".gitignore")
-	if _, err := os.Stat(gitignorePath); err == nil {
-		gitignoreMatcher, _ = ignore.CompileIgnoreFile(gitignorePath)
+	if gitignoreContent, err := filesystem.ReadFile(gitignorePath); err == nil {
+		// Parse gitignore content directly (works with in-memory filesystems)
+		gitignoreMatcher = ignore.CompileIgnoreLines(strings.Split(string(gitignoreContent), "\n")...)
 	}
 
-	// Find all relevant files in workspace
-	err := filepath.Walk(workspaceRoot, func(path string, info os.FileInfo, err error) error {
+	// Find all relevant files in workspace using fs.WalkDir
+	walkErr := fs.WalkDir(filesystem, workspaceRoot, func(path string, d fs.DirEntry, err error) error {
 		if err != nil {
 			return nil // Skip errors and continue
 		}
@@ -314,15 +320,15 @@ func findReferencesInWorkspace(workspaceRoot string, elementName string, openDoc
 		}
 
 		// Skip directories that should be ignored
-		if info.IsDir() {
+		if d.IsDir() {
 			// Always skip .git directory
-			if filepath.Base(path) == ".git" {
-				return filepath.SkipDir
+			if d.Name() == ".git" {
+				return fs.SkipDir
 			}
 
 			// Check gitignore for directories
 			if gitignoreMatcher != nil && gitignoreMatcher.MatchesPath(relPath+"/") {
-				return filepath.SkipDir
+				return fs.SkipDir
 			}
 
 			return nil
@@ -339,34 +345,44 @@ func findReferencesInWorkspace(workspaceRoot string, elementName string, openDoc
 			return nil
 		}
 
+		// Create file URI - ensure path is absolute
+		absPath := path
+		if !filepath.IsAbs(path) {
+			// Relative path - join with workspace root to make absolute
+			absPath = filepath.Join(workspaceRoot, path)
+			// If still relative after joining (e.g., workspace root is "."), prepend / for URI
+			if !filepath.IsAbs(absPath) {
+				absPath = "/" + absPath
+			}
+		}
+		fileURI := "file://" + filepath.ToSlash(absPath)
+
 		// Skip if this file is already open (already searched)
-		fileURI := fmt.Sprintf("file://%s", filepath.ToSlash(path))
 		if openFiles[fileURI] {
 			return nil
 		}
 
 		// Search for references in this file
-		fileLocations := findReferencesInFile(path, fileURI, elementName)
+		fileLocations := findReferencesInFileWithFS(path, fileURI, elementName, filesystem)
 		locations = append(locations, fileLocations...)
 
 		return nil
 	})
 
-	if err != nil {
-		helpers.SafeDebugLog("[REFERENCES] Error walking workspace: %v", err)
+	if walkErr != nil {
+		helpers.SafeDebugLog("[REFERENCES] Error walking workspace: %v", walkErr)
 	}
 
 	helpers.SafeDebugLog("[REFERENCES] Found %d workspace references for %s", len(locations), elementName)
 	return locations
 }
 
-// findReferencesInFile searches for references in a single file on disk using tree-sitter
-func findReferencesInFile(filePath string, fileURI string, elementName string) []protocol.Location {
+// findReferencesInFileWithFS searches for references using a provided filesystem
+func findReferencesInFileWithFS(filePath string, fileURI string, elementName string, filesystem platform.FileSystem) []protocol.Location {
 	var locations []protocol.Location
 
 	// Read file content
-	fs := platform.NewOSFileSystem()
-	content, err := fs.ReadFile(filePath)
+	content, err := filesystem.ReadFile(filePath)
 	if err != nil {
 		helpers.SafeDebugLog("[REFERENCES] Failed to read file %s: %v", filePath, err)
 		return locations
@@ -478,11 +494,32 @@ func findTypeScriptReferencesInContent(content []byte, fileURI string, elementNa
 	defer matcher.Close()
 
 	// Find template literals, then search within them for HTML elements
+	// Track seen templates to avoid duplicates
+	type templateKey struct {
+		uri    string
+		offset uint
+	}
+	seenTemplates := make(map[templateKey]bool)
+
 	for match := range matcher.AllQueryMatches(tree.RootNode(), content) {
 		for _, capture := range match.Captures {
+			captureName := matcher.GetCaptureNameByIndex(capture.Index)
+
+			// Only process template literal captures (skip tag/function captures)
+			if !strings.HasSuffix(captureName, ".literal") && captureName != "innerHTML.template" {
+				continue
+			}
+
 			// Get template content
 			templateContent := content[capture.Node.StartByte():capture.Node.EndByte()]
 			templateOffset := capture.Node.StartByte()
+
+			// Skip if we've already processed this template (same offset)
+			key := templateKey{uri: fileURI, offset: templateOffset}
+			if seenTemplates[key] {
+				continue
+			}
+			seenTemplates[key] = true
 
 			// Parse template content as HTML
 			htmlLocations := findHTMLReferencesInTemplate(templateContent, templateOffset, fileURI, elementName)
