@@ -17,6 +17,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 package lsp
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"os"
@@ -25,15 +26,22 @@ import (
 	"strings"
 	"sync"
 
+	"bennypowers.dev/cem/generate"
+	"bennypowers.dev/cem/internal/logging"
 	"bennypowers.dev/cem/internal/platform"
 	"bennypowers.dev/cem/lsp/helpers"
 	M "bennypowers.dev/cem/manifest"
-	W "bennypowers.dev/cem/workspace"
+	"bennypowers.dev/cem/modulegraph"
+	"bennypowers.dev/cem/queries"
+	"bennypowers.dev/cem/types"
+	"bennypowers.dev/cem/workspace"
+	"gopkg.in/yaml.v3" // Required for parsing pnpm-workspace.yaml files
 )
 
 // ElementDefinition stores a custom element with its source information
 type ElementDefinition struct {
 	CustomElement *M.CustomElement
+	className     string             // Class name from the declaration
 	modulePath    string             // Path from the manifest module
 	Source        *M.SourceReference // Source reference if available
 	packageName   string             // Package name from package.json (if loaded from a package)
@@ -62,6 +70,24 @@ func (ed *ElementDefinition) Element() *M.CustomElement {
 	return ed.CustomElement
 }
 
+// GetModulePath returns the module path (for module graph interface compatibility)
+func (ed *ElementDefinition) GetModulePath() string {
+	return ed.modulePath
+}
+
+// GetTagName returns the tag name
+func (ed *ElementDefinition) GetTagName() string {
+	if ed.CustomElement != nil {
+		return ed.CustomElement.TagName
+	}
+	return ""
+}
+
+// GetClassName returns the class name
+func (ed *ElementDefinition) GetClassName() string {
+	return ed.className
+}
+
 // Registry manages all loaded custom elements manifests and provides
 // fast lookup capabilities for LSP features
 type Registry struct {
@@ -82,17 +108,30 @@ type Registry struct {
 	// File watching
 	fileWatcher platform.FileWatcher
 	watcherMu   sync.RWMutex
-	onReload    func() // Callback when manifests are reloaded
+	watcherDone chan struct{}  // Signal to stop file watching
+	watcherWg   sync.WaitGroup // Wait for watcher goroutine to exit
+	onReload    func()         // Callback when manifests are reloaded
 	// Generate watching for local project
-	generateWatcher *InProcessGenerateWatcher
+	generateWatcher platform.GenerateWatcher
 	generateMu      sync.RWMutex
-	localWorkspace  W.WorkspaceContext // Track the local workspace for generate watching
+	localWorkspace  types.WorkspaceContext // Track the local workspace for generate watching
+	// Module graph for tracking re-export relationships
+	moduleGraph *modulegraph.ModuleGraph
 }
 
 // NewRegistry creates a new empty registry with the given file watcher.
 // For production use, pass platform.NewFSNotifyFileWatcher().
 // For testing, pass platform.NewMockFileWatcher().
 func NewRegistry(fileWatcher platform.FileWatcher) *Registry {
+	// Get QueryManager for dependency injection
+	queryManager, err := queries.GetGlobalQueryManager()
+	if err != nil {
+		// For production, this should not happen, but handle gracefully
+		queryManager = nil
+	}
+
+	moduleGraph := modulegraph.NewModuleGraph(queryManager)
+
 	return &Registry{
 		Elements:             make(map[string]*M.CustomElement),
 		ElementDefinitions:   make(map[string]*ElementDefinition),
@@ -101,6 +140,7 @@ func NewRegistry(fileWatcher platform.FileWatcher) *Registry {
 		ManifestPaths:        make([]string, 0),
 		ManifestPackageNames: make(map[string]string),
 		fileWatcher:          fileWatcher,
+		moduleGraph:          moduleGraph,
 	}
 }
 
@@ -116,8 +156,8 @@ func NewRegistryWithDefaults() (*Registry, error) {
 
 // LoadFromWorkspace loads all available custom elements manifests from
 // the workspace context
-func (r *Registry) LoadFromWorkspace(workspace W.WorkspaceContext) error {
-	helpers.SafeDebugLog("Loading manifests from workspace...")
+func (r *Registry) LoadFromWorkspace(workspace types.WorkspaceContext) error {
+	logging.Info("[REGISTRY] Loading manifests from workspace root: %s", workspace.Root())
 
 	// Clear existing data
 	r.clear()
@@ -127,15 +167,26 @@ func (r *Registry) LoadFromWorkspace(workspace W.WorkspaceContext) error {
 		helpers.SafeDebugLog("Warning: Could not load workspace manifest: %v", err)
 	}
 
-	// 2. Load manifests from node_modules packages
+	// 2. Load manifests from workspace packages (npm/yarn/pnpm workspaces)
+	if err := r.loadWorkspacePackageManifests(workspace); err != nil {
+		helpers.SafeDebugLog("Warning: Could not load workspace package manifests: %v", err)
+	}
+
+	// 3. Load manifests from node_modules packages
 	if err := r.loadNodeModulesManifests(workspace); err != nil {
 		helpers.SafeDebugLog("Warning: Could not load node_modules manifests: %v", err)
 	}
 
-	// 3. Load manifests specified in config
+	// 4. Load manifests specified in config
 	if err := r.loadConfigManifests(workspace); err != nil {
 		helpers.SafeDebugLog("Warning: Could not load config manifests: %v", err)
 	}
+
+	// 5. Initialize module graph for lazy building
+	// Instead of scanning the entire workspace upfront (which was 5000+ files),
+	// we now build the module graph lazily as imports are discovered in open documents.
+	// This provides fast startup while maintaining accurate re-export resolution.
+	r.initializeLazyModuleGraph(workspace)
 
 	helpers.SafeDebugLog("Loaded %d custom elements from %d manifests", len(r.Elements), len(r.Manifests))
 	return nil
@@ -149,6 +200,13 @@ func (r *Registry) clear() {
 	r.Manifests = r.Manifests[:0]
 	r.ManifestPaths = r.ManifestPaths[:0]
 	r.ManifestPackageNames = make(map[string]string)
+	// Get QueryManager for dependency injection
+	queryManager, err := queries.GetGlobalQueryManager()
+	if err != nil {
+		// For production, this should not happen, but handle gracefully
+		queryManager = nil
+	}
+	r.moduleGraph = modulegraph.NewModuleGraph(queryManager)
 }
 
 // clearDataOnly resets the registry data but preserves manifest paths for watching
@@ -160,18 +218,49 @@ func (r *Registry) clearDataOnly() {
 	r.ElementDefinitions = make(map[string]*ElementDefinition)
 	r.attributes = make(map[string]map[string]*M.Attribute)
 	r.Manifests = r.Manifests[:0]
+	// Get QueryManager for dependency injection
+	queryManager, err := queries.GetGlobalQueryManager()
+	if err != nil {
+		// For production, this should not happen, but handle gracefully
+		queryManager = nil
+	}
+	r.moduleGraph = modulegraph.NewModuleGraph(queryManager)
 	// Note: ManifestPaths are preserved for file watching
 }
 
 // loadWorkspaceManifest loads the manifest from the current workspace
-func (r *Registry) loadWorkspaceManifest(workspace W.WorkspaceContext) error {
+func (r *Registry) loadWorkspaceManifest(workspace types.WorkspaceContext) error {
 	helpers.SafeDebugLog("Attempting to load workspace manifest...")
 	helpers.SafeDebugLog("Workspace root: %s", workspace.Root())
 	helpers.SafeDebugLog("Workspace manifest path: %s", workspace.CustomElementsManifestPath())
 
+	// Try to get package name from workspace package.json first
+	var packageName string
+	if packageJSON, err := workspace.PackageJSON(); err == nil && packageJSON != nil {
+		packageName = packageJSON.Name
+		helpers.SafeDebugLog("Package name from workspace package.json: '%s'", packageName)
+	} else {
+		helpers.SafeDebugLog("Could not read workspace package.json: %v", err)
+	}
+
 	pkg, err := workspace.Manifest()
 	if err != nil {
 		helpers.SafeDebugLog("Error loading workspace manifest: %v", err)
+
+		// If manifest loading failed, try to generate it in-memory
+		// This handles both cases:
+		// 1. package.json has "customElements" field but file doesn't exist
+		// 2. package.json has no "customElements" field (RHDS case)
+		helpers.SafeDebugLog("Workspace manifest not available, attempting in-memory generation")
+		if generatedPkg := r.generateInMemoryManifest(workspace.Root(), packageName); generatedPkg != nil {
+			r.addManifest(generatedPkg, packageName)
+			r.localWorkspace = workspace
+			helpers.SafeDebugLog("Successfully generated in-memory manifest for workspace")
+			return nil
+		}
+		helpers.SafeDebugLog("Failed to generate in-memory manifest for workspace")
+
+		// Return original error since in-memory generation also failed
 		return err
 	}
 
@@ -184,15 +273,6 @@ func (r *Registry) loadWorkspaceManifest(workspace W.WorkspaceContext) error {
 					helpers.SafeDebugLog("    Declaration [%d]: %s (tag: %s)", j, customElementDecl.Name, customElementDecl.TagName)
 				}
 			}
-		}
-
-		// Try to get package name from workspace package.json
-		var packageName string
-		if packageJSON, err := workspace.PackageJSON(); err == nil && packageJSON != nil {
-			packageName = packageJSON.Name
-			helpers.SafeDebugLog("Package name from workspace package.json: '%s'", packageName)
-		} else {
-			helpers.SafeDebugLog("Could not read workspace package.json: %v", err)
 		}
 
 		r.addManifest(pkg, packageName)
@@ -210,14 +290,259 @@ func (r *Registry) loadWorkspaceManifest(workspace W.WorkspaceContext) error {
 
 		helpers.SafeDebugLog("Loaded workspace manifest with %d modules", len(pkg.Modules))
 	} else {
-		helpers.SafeDebugLog("Workspace manifest is nil")
+		helpers.SafeDebugLog("Workspace manifest is nil, attempting in-memory generation")
+		// If no manifest but workspace exists, try to generate in-memory
+		if generatedPkg := r.generateInMemoryManifest(workspace.Root(), packageName); generatedPkg != nil {
+			r.addManifest(generatedPkg, packageName)
+			r.localWorkspace = workspace
+			helpers.SafeDebugLog("Successfully generated in-memory manifest for workspace")
+		} else {
+			helpers.SafeDebugLog("Failed to generate in-memory manifest for workspace")
+		}
 	}
 
 	return nil
 }
 
+// loadWorkspacePackageManifests loads manifests from workspace packages (npm/yarn/pnpm workspaces)
+func (r *Registry) loadWorkspacePackageManifests(workspace types.WorkspaceContext) error {
+	helpers.SafeDebugLog("Attempting to load workspace package manifests...")
+
+	// Detect package manager and get workspace configuration
+	workspacePackages, err := r.discoverWorkspacePackages(workspace)
+	if err != nil {
+		return fmt.Errorf("could not discover workspace packages: %w", err)
+	}
+
+	if len(workspacePackages) == 0 {
+		helpers.SafeDebugLog("No workspace packages found")
+		return nil
+	}
+
+	helpers.SafeDebugLog("Found %d workspace packages", len(workspacePackages))
+
+	// Load manifest from each workspace package
+	for _, pkgPath := range workspacePackages {
+		r.loadWorkspacePackage(pkgPath)
+	}
+
+	return nil
+}
+
+// discoverWorkspacePackages discovers workspace packages based on package manager config
+func (r *Registry) discoverWorkspacePackages(workspace types.WorkspaceContext) ([]string, error) {
+	root := workspace.Root()
+	var workspacePatterns []string
+
+	// Detect package manager and get workspace patterns
+	// 1. Try pnpm (pnpm-workspace.yaml)
+	// pnpm uses a separate YAML file for workspace configuration instead of package.json
+	pnpmWorkspaceFile := filepath.Join(root, "pnpm-workspace.yaml")
+	if _, err := os.Stat(pnpmWorkspaceFile); err == nil {
+		patterns, err := r.parsePnpmWorkspace(pnpmWorkspaceFile)
+		if err == nil {
+			workspacePatterns = patterns
+			helpers.SafeDebugLog("Detected pnpm workspace with %d patterns", len(patterns))
+		}
+	}
+
+	// 2. Try npm/yarn (package.json workspaces field)
+	// Only check package.json if we haven't found patterns yet (pnpm takes precedence)
+	if len(workspacePatterns) == 0 {
+		packageJSONPath := filepath.Join(root, "package.json")
+		if _, err := os.Stat(packageJSONPath); err == nil {
+			patterns, err := r.parseNpmYarnWorkspaces(packageJSONPath)
+			if err == nil && len(patterns) > 0 {
+				workspacePatterns = patterns
+				// Determine if npm or yarn based on lock file (for debug logging only)
+				if _, err := os.Stat(filepath.Join(root, "yarn.lock")); err == nil {
+					helpers.SafeDebugLog("Detected yarn workspace with %d patterns", len(patterns))
+				} else {
+					helpers.SafeDebugLog("Detected npm workspace with %d patterns", len(patterns))
+				}
+			}
+		}
+	}
+
+	if len(workspacePatterns) == 0 {
+		return nil, nil
+	}
+
+	// Separate positive and negative patterns
+	var positivePatterns []string
+	var negativePatterns []string
+	for _, pattern := range workspacePatterns {
+		if strings.HasPrefix(pattern, "!") {
+			// Remove the "!" prefix for negation patterns
+			negativePatterns = append(negativePatterns, strings.TrimPrefix(pattern, "!"))
+		} else {
+			positivePatterns = append(positivePatterns, pattern)
+		}
+	}
+
+	// Expand positive glob patterns to find actual workspace package directories
+	var packageDirs []string
+	for _, pattern := range positivePatterns {
+		matches, err := r.expandWorkspacePattern(workspace, pattern)
+		if err != nil {
+			helpers.SafeDebugLog("Warning: Could not expand pattern %s: %v", pattern, err)
+			continue
+		}
+		packageDirs = append(packageDirs, matches...)
+	}
+
+	// Filter out packages matching negation patterns
+	if len(negativePatterns) > 0 {
+		packageDirs = r.filterNegatedPackages(workspace, packageDirs, negativePatterns)
+	}
+
+	helpers.SafeDebugLog("Expanded %d patterns to %d package directories", len(workspacePatterns), len(packageDirs))
+	return packageDirs, nil
+}
+
+// parsePnpmWorkspace parses pnpm-workspace.yaml and returns workspace patterns
+func (r *Registry) parsePnpmWorkspace(path string) ([]string, error) {
+	data, err := os.ReadFile(path)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read pnpm-workspace.yaml: %w", err)
+	}
+
+	var config struct {
+		Packages []string `yaml:"packages"`
+	}
+
+	if err := yaml.Unmarshal(data, &config); err != nil {
+		return nil, fmt.Errorf("failed to parse pnpm-workspace.yaml: %w", err)
+	}
+
+	return config.Packages, nil
+}
+
+// parseNpmYarnWorkspaces parses package.json workspaces field
+func (r *Registry) parseNpmYarnWorkspaces(path string) ([]string, error) {
+	data, err := os.ReadFile(path)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read package.json: %w", err)
+	}
+
+	var pkg struct {
+		Workspaces interface{} `json:"workspaces"`
+	}
+
+	if err := json.Unmarshal(data, &pkg); err != nil {
+		return nil, fmt.Errorf("failed to parse package.json: %w", err)
+	}
+
+	if pkg.Workspaces == nil {
+		return nil, nil
+	}
+
+	// Handle both array format and object format
+	switch w := pkg.Workspaces.(type) {
+	case []interface{}:
+		// Array format: ["packages/*"]
+		patterns := make([]string, 0, len(w))
+		for _, p := range w {
+			if str, ok := p.(string); ok {
+				patterns = append(patterns, str)
+			}
+		}
+		return patterns, nil
+	case map[string]interface{}:
+		// Object format: {"packages": ["packages/*"]}
+		if pkgs, ok := w["packages"].([]interface{}); ok {
+			patterns := make([]string, 0, len(pkgs))
+			for _, p := range pkgs {
+				if str, ok := p.(string); ok {
+					patterns = append(patterns, str)
+				}
+			}
+			return patterns, nil
+		}
+	}
+
+	return nil, nil
+}
+
+// expandWorkspacePattern expands a glob pattern to actual directories using workspace.Glob
+func (r *Registry) expandWorkspacePattern(workspace types.WorkspaceContext, pattern string) ([]string, error) {
+	// Use workspace.Glob which supports ** doublestar patterns
+	matches, err := workspace.Glob(pattern)
+	if err != nil {
+		return nil, fmt.Errorf("failed to expand pattern %s: %w", pattern, err)
+	}
+
+	root := workspace.Root()
+
+	// Filter to only directories that contain package.json
+	var packageDirs []string
+	for _, relPath := range matches {
+		absPath := filepath.Join(root, relPath)
+		info, err := os.Stat(absPath)
+		if err != nil || !info.IsDir() {
+			continue
+		}
+
+		// Check if this directory has a package.json
+		packageJSONPath := filepath.Join(absPath, "package.json")
+		if _, err := os.Stat(packageJSONPath); err == nil {
+			packageDirs = append(packageDirs, absPath)
+		}
+	}
+
+	return packageDirs, nil
+}
+
+// filterNegatedPackages removes packages matching negation patterns
+func (r *Registry) filterNegatedPackages(workspace types.WorkspaceContext, packageDirs []string, negativePatterns []string) []string {
+	root := workspace.Root()
+	var filtered []string
+
+	for _, pkgDir := range packageDirs {
+		// Convert to relative path for pattern matching
+		relPath, err := filepath.Rel(root, pkgDir)
+		if err != nil {
+			// If we can't get relative path, skip this package
+			continue
+		}
+
+		excluded := false
+		for _, negPattern := range negativePatterns {
+			// Try to match the negation pattern
+			matches, err := workspace.Glob(negPattern)
+			if err != nil {
+				continue
+			}
+
+			// Check if this package matches any negation pattern
+			for _, match := range matches {
+				if filepath.Clean(match) == filepath.Clean(relPath) {
+					excluded = true
+					helpers.SafeDebugLog("Excluding package %s (matched negation pattern %s)", relPath, negPattern)
+					break
+				}
+			}
+			if excluded {
+				break
+			}
+		}
+
+		if !excluded {
+			filtered = append(filtered, pkgDir)
+		}
+	}
+
+	return filtered
+}
+
+// loadWorkspacePackage loads a manifest from a single workspace package directory
+func (r *Registry) loadWorkspacePackage(pkgPath string) {
+	helpers.SafeDebugLog("Loading workspace package from: %s", pkgPath)
+	r.loadPackageManifest(pkgPath)
+}
+
 // loadNodeModulesManifests loads manifests from node_modules packages
-func (r *Registry) loadNodeModulesManifests(workspace W.WorkspaceContext) error {
+func (r *Registry) loadNodeModulesManifests(workspace types.WorkspaceContext) error {
 	// Look for node_modules directory
 	nodeModulesPath := filepath.Join(workspace.Root(), "node_modules")
 	entries, err := os.ReadDir(nodeModulesPath)
@@ -253,6 +578,7 @@ func (r *Registry) loadNodeModulesManifests(workspace W.WorkspaceContext) error 
 }
 
 // loadPackageManifest loads a manifest from a specific package directory
+// If the manifest file doesn't exist, it generates it in-memory
 func (r *Registry) loadPackageManifest(packagePath string) {
 	// Read package.json to find customElements field
 	packageJSONPath := filepath.Join(packagePath, "package.json")
@@ -271,15 +597,67 @@ func (r *Registry) loadPackageManifest(packagePath string) {
 		manifestPath = filepath.Join(packagePath, packageJSON.CustomElements)
 	}
 
-	// Load the manifest
+	// Try to load the manifest file
 	if pkg, err := r.loadManifestFileWithPackageName(manifestPath, packageJSON.Name); err == nil {
 		r.addManifest(pkg, packageJSON.Name)
 		helpers.SafeDebugLog("Loaded manifest from %s (%s)", packageJSON.Name, manifestPath)
+		return
+	}
+
+	// If manifest file doesn't exist, generate it in-memory
+	if _, statErr := os.Stat(manifestPath); os.IsNotExist(statErr) {
+		helpers.SafeDebugLog("Manifest file %s doesn't exist, generating in-memory for %s", manifestPath, packageJSON.Name)
+		if pkg := r.generateInMemoryManifest(packagePath, packageJSON.Name); pkg != nil {
+			r.addManifest(pkg, packageJSON.Name)
+			helpers.SafeDebugLog("Generated in-memory manifest for %s with %d modules", packageJSON.Name, len(pkg.Modules))
+		} else {
+			helpers.SafeDebugLog("Failed to generate in-memory manifest for %s", packageJSON.Name)
+		}
+	} else {
+		helpers.SafeDebugLog("Failed to load manifest from %s (%s): file exists but couldn't be parsed", packageJSON.Name, manifestPath)
 	}
 }
 
+// generateInMemoryManifest generates a custom elements manifest in-memory for a package directory
+// This is used when a workspace package declares customElements but the file doesn't exist yet
+func (r *Registry) generateInMemoryManifest(packagePath string, packageName string) *M.Package {
+	logging.Info("[IN-MEMORY] Starting in-memory generation for package '%s' at path: %s", packageName, packagePath)
+
+	// Create a workspace context for the package directory
+	wsCtx := workspace.NewFileSystemWorkspaceContext(packagePath)
+
+	// Initialize the workspace context to load config
+	if err := wsCtx.Init(); err != nil {
+		logging.Warning("[IN-MEMORY] Failed to initialize workspace context for %s: %v", packageName, err)
+		return nil
+	}
+
+	logging.Info("[IN-MEMORY] Successfully initialized workspace context for %s", packageName)
+
+	// Create a generate session
+	session, err := generate.NewGenerateSession(wsCtx)
+	if err != nil {
+		logging.Warning("[IN-MEMORY] Failed to create generate session for %s: %v", packageName, err)
+		return nil
+	}
+	defer session.Close()
+
+	logging.Info("[IN-MEMORY] Created generate session, starting generation for %s...", packageName)
+
+	// Generate the manifest
+	pkg, err := session.GenerateFullManifest(context.Background())
+	if err != nil {
+		logging.Warning("[IN-MEMORY] Failed to generate manifest for %s: %v", packageName, err)
+		return nil
+	}
+
+	logging.Info("[IN-MEMORY] Successfully generated manifest for %s with %d modules", packageName, len(pkg.Modules))
+
+	return pkg
+}
+
 // loadConfigManifests loads manifests specified in the config
-func (r *Registry) loadConfigManifests(workspace W.WorkspaceContext) error {
+func (r *Registry) loadConfigManifests(workspace types.WorkspaceContext) error {
 	// TODO: Add config support for specifying additional manifest paths
 	// This would allow users to specify manifests from non-npm sources
 	return nil
@@ -348,6 +726,9 @@ func (r *Registry) addManifest(manifest *M.Package, packageName string) {
 
 	r.Manifests = append(r.Manifests, manifest)
 
+	// Collect tag names for this manifest to log
+	var tagNames []string
+
 	// Index all custom elements from all modules
 	for _, module := range manifest.Modules {
 		for _, decl := range module.Declarations {
@@ -361,6 +742,7 @@ func (r *Registry) addManifest(manifest *M.Package, packageName string) {
 					// Store the element definition with source information
 					elementDef := &ElementDefinition{
 						CustomElement: element,
+						className:     customElementDecl.Name, // Store the class name from the declaration
 						modulePath:    module.Path,
 						Source:        customElementDecl.Source,
 						packageName:   packageName,
@@ -371,6 +753,9 @@ func (r *Registry) addManifest(manifest *M.Package, packageName string) {
 					}
 					helpers.SafeDebugLog("[REGISTRY] Registering element '%s' with packageName='%s', modulePath='%s'", element.TagName, packageName, module.Path)
 					r.ElementDefinitions[element.TagName] = elementDef
+
+					// Collect tag name for logging
+					tagNames = append(tagNames, element.TagName)
 
 					// Index attributes for this element
 					if element.Attributes != nil {
@@ -385,6 +770,14 @@ func (r *Registry) addManifest(manifest *M.Package, packageName string) {
 			}
 		}
 	}
+
+	// Log the manifest with all its elements
+	displayName := packageName
+	if displayName == "" {
+		displayName = "(unknown)"
+	}
+	elementsList := strings.Join(tagNames, ", ")
+	logging.Info("[REGISTRY] Loaded `%s` elements: %s", displayName, elementsList)
 }
 
 // Element returns the custom element definition for a tag name
@@ -440,15 +833,38 @@ func (r *Registry) ElementDefinition(tagName string) (*ElementDefinition, bool) 
 	r.mu.RLock()
 	defer r.mu.RUnlock()
 
+	// First check manifest-based definitions
 	definition, exists := r.ElementDefinitions[tagName]
-	helpers.SafeDebugLog("[REGISTRY] GetElementDefinition('%s'): exists=%t", tagName, exists)
 	if exists {
+		helpers.SafeDebugLog("[REGISTRY] GetElementDefinition('%s'): found in manifests", tagName)
 		helpers.SafeDebugLog("[REGISTRY] Element '%s' module path: %s", tagName, definition.modulePath)
 		if definition.Source != nil {
 			helpers.SafeDebugLog("[REGISTRY] Element '%s' source href: %s", tagName, definition.Source.Href)
 		}
+		return definition, true
 	}
-	return definition, exists
+
+	// If not found in manifests, check module graph
+	if r.moduleGraph != nil {
+		elementSources := r.moduleGraph.GetElementSources(tagName)
+		if len(elementSources) > 0 {
+			helpers.SafeDebugLog("[REGISTRY] GetElementDefinition('%s'): found in module graph with sources: %v", tagName, elementSources)
+			// Create a minimal ElementDefinition for module graph elements
+			// Use the first source as the primary module path
+			modulePath := elementSources[0]
+			return &ElementDefinition{
+				CustomElement: &M.CustomElement{
+					TagName: tagName,
+				},
+				className:   "", // Class name not available from module graph
+				modulePath:  modulePath,
+				packageName: "", // Package names not available for module graph elements - they come from file scanning, not manifests
+			}, true
+		}
+	}
+
+	helpers.SafeDebugLog("[REGISTRY] GetElementDefinition('%s'): not found", tagName)
+	return nil, false
 }
 
 // AllTagNames returns all registered custom element tag names
@@ -456,10 +872,30 @@ func (r *Registry) AllTagNames() []string {
 	r.mu.RLock()
 	defer r.mu.RUnlock()
 
+	// Collect tags from manifests
 	tags := make([]string, 0, len(r.Elements))
 	for tag := range r.Elements {
 		tags = append(tags, tag)
 	}
+
+	// Also include tags from module graph (for elements discovered from source files)
+	if r.moduleGraph != nil {
+		moduleGraphTags := r.moduleGraph.GetAllTagNames()
+
+		// Use map for O(1) duplicate detection instead of O(n²) nested loops
+		seen := make(map[string]bool, len(tags))
+		for _, tag := range tags {
+			seen[tag] = true
+		}
+
+		for _, tag := range moduleGraphTags {
+			if !seen[tag] {
+				tags = append(tags, tag)
+				seen[tag] = true
+			}
+		}
+	}
+
 	return tags
 }
 
@@ -471,6 +907,7 @@ func (r *Registry) AddManifest(pkg *M.Package) {
 }
 
 // StartFileWatching initializes file watching for manifest changes
+// Thread-safety: All operations in this function are protected by watcherMu.Lock()
 func (r *Registry) StartFileWatching(onReload func()) error {
 	r.watcherMu.Lock()
 	defer r.watcherMu.Unlock()
@@ -479,7 +916,14 @@ func (r *Registry) StartFileWatching(onReload func()) error {
 		return fmt.Errorf("no file watcher configured")
 	}
 
+	// Prevent starting the watcher multiple times
+	if r.watcherDone != nil {
+		return fmt.Errorf("file watcher already running")
+	}
+
 	r.onReload = onReload
+	// PROTECTED BY LOCK: Channel initialization is thread-safe with StopFileWatching
+	r.watcherDone = make(chan struct{})
 
 	// Add all known manifest paths to the watcher
 	for _, path := range r.ManifestPaths {
@@ -491,12 +935,16 @@ func (r *Registry) StartFileWatching(onReload func()) error {
 	}
 
 	// Start watching in a goroutine
-	go r.watchFiles()
+	r.watcherWg.Go(func() {
+		r.watchFiles()
+		helpers.SafeDebugLog("File watcher goroutine exiting")
+	})
 
 	return nil
 }
 
 // StopFileWatching stops file watching
+// Thread-safety: All operations in this function are protected by watcherMu.Lock()
 func (r *Registry) StopFileWatching() error {
 	r.watcherMu.Lock()
 	defer r.watcherMu.Unlock()
@@ -505,6 +953,25 @@ func (r *Registry) StopFileWatching() error {
 		return nil
 	}
 
+	// PROTECTED BY LOCK: This nil check and channel close are atomic with respect
+	// to StartFileWatching, preventing races between concurrent Start/Stop calls
+	if r.watcherDone != nil {
+		close(r.watcherDone)
+		// Closing is sufficient for shutdown signaling
+	}
+
+	// Release lock before waiting to avoid deadlock
+	r.watcherMu.Unlock()
+
+	// Wait for goroutine to exit
+	r.watcherWg.Wait()
+	helpers.SafeDebugLog("File watcher goroutine has exited")
+
+	// Reacquire lock for final cleanup
+	r.watcherMu.Lock()
+
+	// Now it's safe to set watcherDone to nil - goroutine has exited
+	r.watcherDone = nil
 	err := r.fileWatcher.Close()
 	r.onReload = nil
 	return err
@@ -513,6 +980,9 @@ func (r *Registry) StopFileWatching() error {
 // watchFiles handles file system events in a goroutine
 func (r *Registry) watchFiles() {
 	// Get channel references under lock to avoid race condition
+	// We copy the channel references once at startup to avoid holding the lock
+	// during the entire watch loop. The channels are immutable once copied,
+	// so this is safe even if StopFileWatching is called concurrently.
 	r.watcherMu.RLock()
 	if r.fileWatcher == nil {
 		r.watcherMu.RUnlock()
@@ -520,8 +990,11 @@ func (r *Registry) watchFiles() {
 	}
 	events := r.fileWatcher.Events()
 	errors := r.fileWatcher.Errors()
+	done := r.watcherDone
 	r.watcherMu.RUnlock()
 
+	// If done is nil, the select will simply ignore that case (nil channels never trigger)
+	// This shouldn't happen in practice, but is safe if it does
 	for {
 		select {
 		case event, ok := <-events:
@@ -534,6 +1007,10 @@ func (r *Registry) watchFiles() {
 				return
 			}
 			helpers.SafeDebugLog("File watcher error: %v", err)
+		case <-done:
+			// This case triggers when StopFileWatching closes the done channel
+			helpers.SafeDebugLog("File watcher stopped due to shutdown signal")
+			return
 		}
 	}
 }
@@ -558,7 +1035,7 @@ func (r *Registry) handleFileChange(event platform.FileWatchEvent) {
 	r.watcherMu.RLock()
 	callback := r.onReload
 	r.watcherMu.RUnlock()
-	
+
 	if callback != nil {
 		callback()
 	}
@@ -614,9 +1091,17 @@ func (r *Registry) StartGenerateWatcher() error {
 	r.generateMu.Lock()
 	defer r.generateMu.Unlock()
 
-	// Only start if we have a local workspace and no watcher is running
-	if r.localWorkspace == nil || r.generateWatcher != nil {
+	// Only start if we have a local workspace
+	if r.localWorkspace == nil {
 		return nil
+	}
+
+	// If a generate watcher is already set (e.g., for testing), just start it
+	if r.generateWatcher != nil {
+		if r.generateWatcher.IsRunning() {
+			return nil // Already running
+		}
+		return r.generateWatcher.Start()
 	}
 
 	// Get the workspace root directory
@@ -645,10 +1130,10 @@ func (r *Registry) StartGenerateWatcher() error {
 		for _, module := range pkg.Modules {
 			for _, decl := range module.Declarations {
 				if customElement, ok := decl.(*M.CustomElementDeclaration); ok {
-					if customElement.CustomElement.TagName == "test-button" {
-						for _, attr := range customElement.CustomElement.Attributes {
+					if customElement.TagName == "test-button" {
+						for _, attr := range customElement.Attributes {
 							if attr.Name == "variant" && attr.Type != nil {
-								fmt.Fprintf(os.Stderr, "[DEBUG] Received manifest - variant type: '%s'\n", attr.Type.Text)
+								helpers.SafeDebugLog("Received manifest - variant type: '%s'", attr.Type.Text)
 							}
 						}
 					}
@@ -772,4 +1257,222 @@ func (r *Registry) addManifestPath(path string) {
 		}
 	}
 	r.watcherMu.RUnlock()
+}
+
+// GetModuleGraph returns the module graph for re-export analysis
+func (r *Registry) GetModuleGraph() *modulegraph.ModuleGraph {
+	return r.moduleGraph
+}
+
+// GetFileWatcher returns the file watcher for testing purposes
+func (r *Registry) GetFileWatcher() platform.FileWatcher {
+	r.watcherMu.RLock()
+	defer r.watcherMu.RUnlock()
+	return r.fileWatcher
+}
+
+// SetGenerateWatcher sets a custom generate watcher for testing purposes
+func (r *Registry) SetGenerateWatcher(watcher platform.GenerateWatcher) {
+	r.generateMu.Lock()
+	defer r.generateMu.Unlock()
+	r.generateWatcher = watcher
+}
+
+// RegistryManifestResolver implements ManifestResolver using the registry's manifest data
+type RegistryManifestResolver struct {
+	registry *Registry
+}
+
+// NewRegistryManifestResolver creates a new manifest resolver using the registry
+func NewRegistryManifestResolver(registry *Registry) *RegistryManifestResolver {
+	return &RegistryManifestResolver{registry: registry}
+}
+
+// FindManifestModulesForImportPath finds manifest modules that match an import path
+func (r *RegistryManifestResolver) FindManifestModulesForImportPath(importPath string) []string {
+	if r.registry == nil {
+		return nil
+	}
+
+	var matchingModules []string
+
+	r.registry.mu.RLock()
+	defer r.registry.mu.RUnlock()
+
+	// Search through all element definitions to find matches
+	for _, elementDef := range r.registry.ElementDefinitions {
+		modulePath := elementDef.GetModulePath()
+		if modulePath == "" {
+			continue
+		}
+
+		// Use the same path matching logic as tagDiagnostics.go
+		if r.pathsMatch(importPath, modulePath) {
+			// Avoid duplicates
+			found := false
+			for _, existing := range matchingModules {
+				if existing == modulePath {
+					found = true
+					break
+				}
+			}
+			if !found {
+				matchingModules = append(matchingModules, modulePath)
+			}
+		}
+	}
+
+	return matchingModules
+}
+
+// GetManifestModulePath converts a file path to its corresponding manifest module path
+func (r *RegistryManifestResolver) GetManifestModulePath(filePath string) string {
+	if r.registry == nil {
+		return ""
+	}
+
+	r.registry.mu.RLock()
+	defer r.registry.mu.RUnlock()
+
+	// Search through all element definitions to find one that matches this file path
+	// This is a reverse lookup from file path to manifest module path
+	for _, elementDef := range r.registry.ElementDefinitions {
+		modulePath := elementDef.GetModulePath()
+		if modulePath == "" {
+			continue
+		}
+
+		// Try direct path matching first
+		if r.pathsMatch(filePath, modulePath) {
+			return modulePath
+		}
+
+		// For Red Hat Design System pattern: convert TypeScript source paths to manifest paths
+		// e.g., "elements/rh-tabs/rh-tab-panel.ts" -> "rh-tabs/rh-tab-panel.js"
+		if strings.HasSuffix(filePath, ".ts") {
+			// Convert .ts to .js for comparison
+			jsFilePath := strings.TrimSuffix(filePath, ".ts") + ".js"
+			if r.pathsMatch(jsFilePath, modulePath) {
+				return modulePath
+			}
+
+			// Also try without elements/ prefix
+			// "elements/rh-tabs/rh-tab-panel.ts" -> "rh-tabs/rh-tab-panel.js"
+			if strings.HasPrefix(filePath, "elements/") {
+				relativeJsPath := strings.TrimPrefix(jsFilePath, "elements/")
+				if r.pathsMatch(relativeJsPath, modulePath) {
+					return modulePath
+				}
+			}
+		}
+	}
+
+	return ""
+}
+
+// GetElementsFromManifestModule returns all custom element tag names available from a manifest module
+func (r *RegistryManifestResolver) GetElementsFromManifestModule(manifestModulePath string) []string {
+	if r.registry == nil {
+		return nil
+	}
+
+	var elements []string
+
+	r.registry.mu.RLock()
+	defer r.registry.mu.RUnlock()
+
+	// Search through all element definitions to find those from this manifest module
+	for tagName, elementDef := range r.registry.ElementDefinitions {
+		modulePath := elementDef.GetModulePath()
+		if modulePath == manifestModulePath {
+			elements = append(elements, tagName)
+		}
+	}
+
+	return elements
+}
+
+// pathsMatch checks if an import path matches an element source path
+// This duplicates the logic from tagDiagnostics.go for now, but should ideally be unified
+func (r *RegistryManifestResolver) pathsMatch(importPath, elementSource string) bool {
+	// Direct match first (for exact package imports)
+	if importPath == elementSource {
+		return true
+	}
+
+	// Normalize paths for comparison
+	normalizedImport := r.normalizePath(importPath)
+	normalizedSource := r.normalizePath(elementSource)
+
+	// Direct match on normalized paths
+	if normalizedImport == normalizedSource {
+		return true
+	}
+
+	// Check if import path ends with the element source (relative imports)
+	if strings.HasSuffix(importPath, elementSource) {
+		return true
+	}
+
+	// Check if element source ends with import path (package imports)
+	if strings.HasSuffix(elementSource, importPath) {
+		return true
+	}
+
+	// Extract just the filename and compare
+	importFile := filepath.Base(importPath)
+	elementFile := filepath.Base(elementSource)
+	return importFile == elementFile
+}
+
+// normalizePath normalizes a file path for comparison
+// This duplicates the logic from tagDiagnostics.go for now, but should ideally be unified
+func (r *RegistryManifestResolver) normalizePath(path string) string {
+	// Remove common prefixes/suffixes
+	path = strings.TrimPrefix(path, "./")
+	path = strings.TrimPrefix(path, "../")
+	path = strings.TrimPrefix(path, "/")
+
+	// Handle npm package paths like @rhds/elements/rh-card/rh-card.js
+	// vs manifest paths like ./dist/rh-card.js
+	if strings.Contains(path, "/") {
+		// Keep the last two segments for better matching
+		parts := strings.Split(path, "/")
+		if len(parts) >= 2 {
+			return strings.Join(parts[len(parts)-2:], "/")
+		}
+	}
+
+	return path
+}
+
+// initializeLazyModuleGraph initializes the module graph with manifest data but defers file scanning
+func (r *Registry) initializeLazyModuleGraph(workspace types.WorkspaceContext) {
+	if workspace == nil {
+		return
+	}
+
+	workspaceRoot := workspace.Root()
+	if workspaceRoot == "" {
+		return
+	}
+
+	helpers.SafeDebugLog("[REGISTRY] Initializing lazy module graph for workspace: %s", workspaceRoot)
+
+	// Create manifest resolver and update module graph to use it
+	manifestResolver := NewRegistryManifestResolver(r)
+	r.moduleGraph.SetManifestResolver(manifestResolver)
+
+	// Populate the module graph with custom element definitions from manifests
+	// This is fast since it only uses already-loaded manifest data
+	elementMap := make(map[string]interface{})
+	for tagName, elementDef := range r.ElementDefinitions {
+		elementMap[tagName] = elementDef
+	}
+	r.moduleGraph.PopulateFromManifests(elementMap)
+
+	// Store workspace root for lazy file parsing
+	r.moduleGraph.SetWorkspaceRoot(workspaceRoot)
+
+	helpers.SafeDebugLog("[REGISTRY] Module graph initialized with %d manifest elements and manifest resolver, ready for lazy building", len(elementMap))
 }
