@@ -1,13 +1,32 @@
+/*
+Copyright © 2025 Benny Powers <web@bennypowers.com>
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
 package lsp_test
 
 import (
 	"encoding/json"
 	"os"
 	"path/filepath"
+	"strings"
 	"sync"
 	"testing"
+	"testing/synctest"
 	"time"
 
+	"bennypowers.dev/cem/internal/platform"
 	"bennypowers.dev/cem/lsp"
 	"bennypowers.dev/cem/lsp/testhelpers"
 	W "bennypowers.dev/cem/workspace"
@@ -19,7 +38,7 @@ func TestManifestFileWatchingIntegration(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to create temp dir: %v", err)
 	}
-	defer os.RemoveAll(tempDir)
+	defer func() { _ = os.RemoveAll(tempDir) }()
 
 	// Copy initial manifest fixture to temp directory
 	fixturePath := filepath.Join("test", "fixtures", "file-watch-integration", "initial-manifest.json")
@@ -63,19 +82,9 @@ func TestManifestFileWatchingIntegration(t *testing.T) {
 	})
 
 	t.Run("file watching triggers reload", func(t *testing.T) {
-		// TODO(Go 1.25): Replace CI skip with testing/synctest virtual time
-		// and testing/fstest.MapFS for deterministic file watching tests.
-		// This will eliminate timing dependencies and enable instant test execution.
-		// See: https://tip.golang.org/doc/go1.25#testing
-		if os.Getenv("CI") == "true" {
-			t.Skip("Skipping file watching integration test in CI environment - timing sensitive until Go 1.25 provides testing/synctest")
-		}
-
-		// Create fresh registry for this test
-		testRegistry, err := lsp.NewRegistryWithDefaults()
-		if err != nil {
-			t.Fatalf("Failed to create registry: %v", err)
-		}
+		// Create registry with MockFileWatcher for testing
+		mockFileWatcher := platform.NewMockFileWatcher()
+		testRegistry := lsp.NewRegistry(mockFileWatcher)
 		err = testRegistry.LoadFromWorkspace(workspace)
 		if err != nil {
 			t.Fatalf("Failed to load from workspace: %v", err)
@@ -84,7 +93,7 @@ func TestManifestFileWatchingIntegration(t *testing.T) {
 		var reloadCount int
 		var reloadMu sync.Mutex
 
-		// Set up file watching
+		// Set up file watching with reload callback
 		onReload := func() {
 			reloadMu.Lock()
 			defer reloadMu.Unlock()
@@ -96,14 +105,16 @@ func TestManifestFileWatchingIntegration(t *testing.T) {
 			if err != nil {
 				return
 			}
-			testRegistry.LoadFromWorkspace(freshWorkspace)
+			if err := testRegistry.LoadFromWorkspace(freshWorkspace); err != nil {
+				return // If loading fails, callback won't work
+			}
 		}
 
 		err = testRegistry.StartFileWatching(onReload)
 		if err != nil {
 			t.Fatalf("Failed to start file watching: %v", err)
 		}
-		defer testRegistry.StopFileWatching()
+		defer func() { _ = testRegistry.StopFileWatching() }()
 
 		// Copy updated manifest fixture to trigger file change
 		updatedFixturePath := filepath.Join("test", "fixtures", "file-watch-integration", "updated-manifest.json")
@@ -112,85 +123,112 @@ func TestManifestFileWatchingIntegration(t *testing.T) {
 			t.Fatalf("Failed to copy updated manifest fixture: %v", err)
 		}
 
-		// Wait for file system event to be processed
-		timeout := time.After(3 * time.Second)
-		for {
-			select {
-			case <-timeout:
-				t.Fatal("Timeout waiting for file change to be detected")
-			default:
-				reloadMu.Lock()
-				count := reloadCount
-				reloadMu.Unlock()
-
-				if count > 0 {
-					// Reload was triggered, verify the registry was updated
-					if len(testRegistry.Elements) == 2 {
-						if _, exists := testRegistry.Elements["new-element"]; exists {
-							return // Success!
-						}
-					}
-				}
-				time.Sleep(100 * time.Millisecond)
-			}
-		}
-	})
-
-	t.Run("multiple file changes", func(t *testing.T) {
-		// Create fresh registry for this test
-		testRegistry, err := lsp.NewRegistryWithDefaults()
+		// Get the absolute path to match what the registry tracks
+		absManifestPath, err := filepath.Abs(manifestPath)
 		if err != nil {
-			t.Fatalf("Failed to create registry: %v", err)
-		}
-		err = testRegistry.LoadFromWorkspace(workspace)
-		if err != nil {
-			t.Fatalf("Failed to load from workspace: %v", err)
+			t.Fatalf("Failed to get absolute path: %v", err)
 		}
 
-		var reloadCount int
-		var reloadMu sync.Mutex
+		// Trigger event with absolute path
+		mockFileWatcher.TriggerEvent(absManifestPath, platform.Write)
 
-		onReload := func() {
-			reloadMu.Lock()
-			defer reloadMu.Unlock()
-			reloadCount++
-			// Create fresh workspace to avoid caching issues
-			freshWorkspace := W.NewFileSystemWorkspaceContext(tempDir)
-			err := freshWorkspace.Init()
-			if err != nil {
-				return
-			}
-			testRegistry.LoadFromWorkspace(freshWorkspace)
-		}
+		// Give time for the event to be processed (using real time)
+		time.Sleep(10 * time.Millisecond)
 
-		err = testRegistry.StartFileWatching(onReload)
-		if err != nil {
-			t.Fatalf("Failed to start file watching: %v", err)
-		}
-		defer testRegistry.StopFileWatching()
-
-		// Make multiple rapid changes by copying fixtures alternately
-		fixtures := []string{"initial-manifest.json", "updated-manifest.json"}
-		for i := 0; i < 3; i++ {
-			fixtureName := fixtures[i%2]
-			fixturePath := filepath.Join("test", "fixtures", "file-watch-integration", fixtureName)
-			err = testhelpers.CopyFile(fixturePath, manifestPath)
-			if err != nil {
-				t.Fatalf("Failed to copy manifest fixture %s: %v", fixtureName, err)
-			}
-			time.Sleep(200 * time.Millisecond)
-		}
-
-		// Wait for changes to be processed
-		time.Sleep(1 * time.Second)
-
+		// Verify reload was triggered
 		reloadMu.Lock()
 		count := reloadCount
 		reloadMu.Unlock()
 
 		if count == 0 {
-			t.Error("Expected at least one reload, got none")
+			t.Fatal("Reload should have been triggered by file watcher event")
 		}
+
+		// Verify the registry was updated with the new manifest content
+		if len(testRegistry.Elements) == 2 {
+			if _, exists := testRegistry.Elements["new-element"]; exists {
+				return // Success!
+			}
+		}
+		t.Errorf("Registry not properly updated after file change. Elements: %d", len(testRegistry.Elements))
+	})
+
+	t.Run("multiple file changes", func(t *testing.T) {
+		synctest.Test(t, func(t *testing.T) {
+			// Create registry with MockFileWatcher for testing
+			mockFileWatcher := platform.NewMockFileWatcher()
+			testRegistry := lsp.NewRegistry(mockFileWatcher)
+			err = testRegistry.LoadFromWorkspace(workspace)
+			if err != nil {
+				t.Fatalf("Failed to load from workspace: %v", err)
+			}
+
+			var reloadCount int
+			var reloadMu sync.Mutex
+
+			onReload := func() {
+				reloadMu.Lock()
+				defer reloadMu.Unlock()
+				reloadCount++
+				// Create fresh workspace to avoid caching issues
+				freshWorkspace := W.NewFileSystemWorkspaceContext(tempDir)
+				err := freshWorkspace.Init()
+				if err != nil {
+					return
+				}
+				if err := testRegistry.LoadFromWorkspace(freshWorkspace); err != nil {
+					return // If loading fails, callback won't work
+				}
+			}
+
+			err = testRegistry.StartFileWatching(onReload)
+			if err != nil {
+				t.Fatalf("Failed to start file watching: %v", err)
+			}
+			defer func() { _ = testRegistry.StopFileWatching() }()
+
+			// Use the MockFileWatcher for triggering events
+
+			// Make multiple rapid changes by copying fixtures alternately
+			fixtures := []string{"initial-manifest.json", "updated-manifest.json"}
+
+			// Get absolute path for consistent triggering
+			absManifestPath, err := filepath.Abs(manifestPath)
+			if err != nil {
+				t.Fatalf("Failed to get absolute path: %v", err)
+			}
+
+			for i := 0; i < 3; i++ {
+				fixtureName := fixtures[i%2]
+				fixturePath := filepath.Join("test", "fixtures", "file-watch-integration", fixtureName)
+				err = testhelpers.CopyFile(fixturePath, manifestPath)
+				if err != nil {
+					t.Fatalf("Failed to copy manifest fixture %s: %v", fixtureName, err)
+				}
+				// Trigger file change event via MockFileWatcher - instant with virtual time
+				mockFileWatcher.TriggerEvent(absManifestPath, platform.Write)
+
+				// Yield to allow the watchFiles goroutine to process the event
+				var quickWait sync.WaitGroup
+				quickWait.Add(1)
+				go func() {
+					defer quickWait.Done()
+				}()
+				quickWait.Wait()
+			}
+
+			// Verify multiple changes were processed - instant with virtual time
+			reloadMu.Lock()
+			count := reloadCount
+			reloadMu.Unlock()
+
+			if count == 0 {
+				t.Error("Expected at least one reload, got none")
+			}
+			if count != 3 {
+				t.Logf("Expected 3 reloads, got %d", count)
+			}
+		})
 	})
 }
 
@@ -200,7 +238,7 @@ func TestPackageJSONWatching(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to create temp dir: %v", err)
 	}
-	defer os.RemoveAll(tempDir)
+	defer func() { _ = os.RemoveAll(tempDir) }()
 
 	// Copy manifest to the referenced location
 	manifestPath := filepath.Join(tempDir, "custom-elements.json")
@@ -264,6 +302,22 @@ func TestPackageJSONWatching(t *testing.T) {
 	})
 
 	t.Run("package.json changes trigger reload", func(t *testing.T) {
+		// Create registry with MockFileWatcher for testing
+		mockFileWatcher := platform.NewMockFileWatcher()
+		testRegistry := lsp.NewRegistry(mockFileWatcher)
+
+		// Load workspace to populate manifest paths
+		workspace := W.NewFileSystemWorkspaceContext(tempDir)
+		err = workspace.Init()
+		if err != nil {
+			t.Fatalf("Failed to initialize workspace: %v", err)
+		}
+		err = testRegistry.LoadFromWorkspace(workspace)
+		if err != nil {
+			t.Fatalf("Failed to load from workspace: %v", err)
+		}
+
+		// Use a simple mutex approach like the other tests
 		var reloadTriggered bool
 		var reloadMu sync.Mutex
 
@@ -273,11 +327,11 @@ func TestPackageJSONWatching(t *testing.T) {
 			reloadTriggered = true
 		}
 
-		err := registry.StartFileWatching(onReload)
+		err = testRegistry.StartFileWatching(onReload)
 		if err != nil {
 			t.Fatalf("Failed to start file watching: %v", err)
 		}
-		defer registry.StopFileWatching()
+		defer func() { _ = testRegistry.StopFileWatching() }()
 
 		// Update package.json by reading, modifying, and writing back
 		var packageJSON map[string]interface{}
@@ -297,22 +351,25 @@ func TestPackageJSONWatching(t *testing.T) {
 			t.Fatalf("Failed to update package.json: %v", err)
 		}
 
-		// Wait for file change detection
-		timeout := time.After(2 * time.Second)
-		for {
-			select {
-			case <-timeout:
-				t.Fatal("Timeout waiting for package.json change to be detected")
-			default:
-				reloadMu.Lock()
-				triggered := reloadTriggered
-				reloadMu.Unlock()
+		// Trigger file change event via MockFileWatcher
+		// Get absolute path for consistent triggering
+		absPackagePath, err := filepath.Abs(packagePath)
+		if err != nil {
+			t.Fatalf("Failed to get absolute path: %v", err)
+		}
 
-				if triggered {
-					return // Success!
-				}
-				time.Sleep(100 * time.Millisecond)
-			}
+		mockFileWatcher.TriggerEvent(absPackagePath, platform.Write)
+
+		// Give time for the event to be processed (using real time)
+		time.Sleep(10 * time.Millisecond)
+
+		// Verify reload was triggered
+		reloadMu.Lock()
+		triggered := reloadTriggered
+		reloadMu.Unlock()
+
+		if !triggered {
+			t.Fatal("package.json change should have triggered reload")
 		}
 	})
 }
@@ -329,7 +386,7 @@ func TestFileWatchingErrorHandling(t *testing.T) {
 		if err != nil {
 			t.Errorf("Expected StartFileWatching to succeed with no files, got error: %v", err)
 		}
-		defer registry.StopFileWatching()
+		defer func() { _ = registry.StopFileWatching() }()
 	})
 
 	t.Run("stop watching when not started", func(t *testing.T) {
@@ -341,17 +398,25 @@ func TestFileWatchingErrorHandling(t *testing.T) {
 	})
 
 	t.Run("double start watching", func(t *testing.T) {
-		err1 := registry.StartFileWatching(func() {})
-		err2 := registry.StartFileWatching(func() {})
+		// Ensure clean state - stop any existing watcher first
+		_ = registry.StopFileWatching()
 
+		// First start should succeed
+		err1 := registry.StartFileWatching(func() {})
 		if err1 != nil {
 			t.Errorf("First StartFileWatching failed: %v", err1)
 		}
-		if err2 != nil {
-			t.Errorf("Second StartFileWatching failed: %v", err2)
+
+		// Second start should fail with "already running" error
+		err2 := registry.StartFileWatching(func() {})
+		if err2 == nil {
+			t.Errorf("Second StartFileWatching should have failed with 'already running' error, but succeeded")
+		} else if !strings.Contains(err2.Error(), "already running") {
+			t.Errorf("Expected 'already running' error, got: %v", err2)
 		}
 
-		defer registry.StopFileWatching()
+		// Cleanup
+		defer func() { _ = registry.StopFileWatching() }()
 	})
 
 	t.Run("watch nonexistent file", func(t *testing.T) {
@@ -367,6 +432,6 @@ func TestFileWatchingErrorHandling(t *testing.T) {
 		if err != nil {
 			t.Errorf("Expected StartFileWatching to handle nonexistent files gracefully, got error: %v", err)
 		}
-		defer testRegistry.StopFileWatching()
+		defer func() { _ = testRegistry.StopFileWatching() }()
 	})
 }
