@@ -32,13 +32,14 @@ import (
 
 	C "bennypowers.dev/cem/cmd/config"
 	M "bennypowers.dev/cem/manifest"
+	"bennypowers.dev/cem/types"
 	"github.com/adrg/xdg"
 	"github.com/bmatcuk/doublestar"
 	"github.com/pterm/pterm"
 	"gopkg.in/yaml.v3"
 )
 
-var _ WorkspaceContext = (*RemoteWorkspaceContext)(nil)
+var _ types.WorkspaceContext = (*RemoteWorkspaceContext)(nil)
 
 // RemoteWorkspaceContext implements WorkspaceContext for remote packages.
 type RemoteWorkspaceContext struct {
@@ -50,10 +51,14 @@ type RemoteWorkspaceContext struct {
 	packageJSON        *M.PackageJSON
 	customElementsPath string
 	spinner            *pterm.SpinnerPrinter
+	designTokensCache  types.DesignTokensCache
 }
 
 func NewRemoteWorkspaceContext(spec string) *RemoteWorkspaceContext {
-	return &RemoteWorkspaceContext{spec: spec}
+	return &RemoteWorkspaceContext{
+		spec:              spec,
+		designTokensCache: NewDesignTokensCache(nil), // Remote contexts don't need design tokens loading
+	}
 }
 
 func (c *RemoteWorkspaceContext) Init() error {
@@ -178,7 +183,7 @@ func (c *RemoteWorkspaceContext) fetchFromNpm(name, version string) error {
 	if err != nil {
 		return err
 	}
-	defer resp.Body.Close()
+	defer func() { _ = resp.Body.Close() }()
 	type npmMeta struct {
 		Versions map[string]struct {
 			Dist struct{ Tarball string }
@@ -193,7 +198,7 @@ func (c *RemoteWorkspaceContext) fetchFromNpm(name, version string) error {
 	if err != nil {
 		return err
 	}
-	defer resp.Body.Close()
+	defer func() { _ = resp.Body.Close() }()
 	err = extractFilesFromTarGz(resp.Body, c.cacheDir, []string{"package.json", "custom-elements.json"})
 	if err != nil {
 		return err
@@ -226,7 +231,7 @@ func (c *RemoteWorkspaceContext) Config() (*C.CemConfig, error) {
 	if err != nil {
 		return nil, err
 	}
-	defer f.Close()
+	defer func() { _ = f.Close() }()
 	var cfg C.CemConfig
 	if err := yaml.NewDecoder(f).Decode(&cfg); err != nil {
 		return nil, err
@@ -243,7 +248,7 @@ func (c *RemoteWorkspaceContext) Manifest() (*M.Package, error) {
 		manifestPath := filepath.Join(c.cacheDir, pkg.CustomElements)
 		rc, err := os.Open(manifestPath)
 		if err == nil {
-			defer rc.Close()
+			defer func() { _ = rc.Close() }()
 			return decodeJSON[M.Package](rc)
 		}
 	}
@@ -255,7 +260,7 @@ func (c *RemoteWorkspaceContext) PackageJSON() (*M.PackageJSON, error) {
 	if err != nil {
 		return nil, fmt.Errorf("could not open package.json in remote workspace: %w", err)
 	}
-	defer rc.Close()
+	defer func() { _ = rc.Close() }()
 	return decodeJSON[M.PackageJSON](rc)
 }
 
@@ -323,6 +328,11 @@ func (c *RemoteWorkspaceContext) ResolveModuleDependency(modulePath, dependencyP
 	return filepath.Clean(resolved), nil
 }
 
+// DesignTokensCache returns the design tokens cache for this workspace
+func (c *RemoteWorkspaceContext) DesignTokensCache() types.DesignTokensCache {
+	return c.designTokensCache
+}
+
 func setupTempdirFromCache(cacheDir string) (string, error) {
 	tempdir, err := os.MkdirTemp("", "cem-remote-*")
 	if err != nil {
@@ -335,7 +345,7 @@ func setupTempdirFromCache(cacheDir string) (string, error) {
 	for _, f := range files {
 		src := filepath.Join(cacheDir, f.Name())
 		dst := filepath.Join(tempdir, f.Name())
-		copyFile(src, dst)
+		_, _ = copyFile(src, dst)
 	}
 	return tempdir, nil
 }
@@ -346,7 +356,7 @@ func (c *RemoteWorkspaceContext) fetch(url, dest string) error {
 	if err != nil {
 		return err
 	}
-	defer resp.Body.Close()
+	defer func() { _ = resp.Body.Close() }()
 
 	if resp.StatusCode == http.StatusNotFound {
 		if strings.HasSuffix(url, "package.json") {
@@ -367,7 +377,7 @@ func (c *RemoteWorkspaceContext) fetch(url, dest string) error {
 	if err != nil {
 		return err
 	}
-	defer out.Close()
+	defer func() { _ = out.Close() }()
 	_, err = io.Copy(out, resp.Body)
 	return err
 }
@@ -377,7 +387,7 @@ func extractFilesFromTarGz(r io.Reader, dest string, wanted []string) error {
 	if err != nil {
 		return err
 	}
-	defer gzr.Close()
+	defer func() { _ = gzr.Close() }()
 	tr := tar.NewReader(gzr)
 	for {
 		hdr, err := tr.Next()
@@ -394,10 +404,10 @@ func extractFilesFromTarGz(r io.Reader, dest string, wanted []string) error {
 					return err
 				}
 				if _, err := io.Copy(out, tr); err != nil {
-					out.Close()
+					_ = out.Close()
 					return err
 				}
-				out.Close()
+				_ = out.Close()
 			}
 		}
 	}
